% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{addHotspotToVision}
\alias{addHotspotToVision}
\title{Add HS python obj to vision OBJECT}
\usage{
addHotspotToVision(object, hs)
}
\arguments{
\item{object}{Vision object}

\item{hs}{python hs object}
}
\value{
Vision object with hs populated
}
\description{
Add HS python obj to vision OBJECT
}
