% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{addLatentSpace}
\alias{addLatentSpace}
\title{Add a latent space computed using an external method}
\usage{
addLatentSpace(object, coordinates, name)
}
\arguments{
\item{object}{the VISION object for which compute the latent space}

\item{coordinates}{matrix with latent space coordinates (cells x dimensions)}

\item{name}{a label for the latent space (e.g. "PCA" or "scVI")}
}
\value{
the VISION with @latentSpace slot populated
}
\description{
Add a latent space computed using an external method
}
