% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\docType{methods}
\name{addProjection,FastProject-method}
\alias{addProjection,FastProject-method}
\title{Add a set of projection coordinates}
\usage{
\S4method{addProjection}{FastProject}(object, name, coordinates)
}
\arguments{
\item{object}{FastProject object}

\item{name}{Name of the projection}

\item{coordinates}{numeric matrix or data.frame. Coordinates of each
sample in the projection (NUM_SAMPLES x NUM_COMPONENTS)}
}
\value{
FastProject object
}
\description{
Add a set of projection coordinates
}
