% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{addProjection,Vision-method}
\alias{addProjection,Vision-method}
\title{Add a set of projection coordinates}
\usage{
\S4method{addProjection}{Vision}(object, name, coordinates)
}
\arguments{
\item{object}{VISION object}

\item{name}{Name of the projection}

\item{coordinates}{numeric matrix or data.frame. Coordinates of each
sample in the projection (NUM_SAMPLES x NUM_COMPONENTS)}
}
\value{
VISION object
}
\description{
Add a set of projection coordinates
}
