% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\docType{methods}
\name{addSigData,Signature-method}
\alias{addSigData,Signature-method}
\title{Add signature data to the sigDict of this Signature Object}
\usage{
\S4method{addSigData}{Signature}(object, data)
}
\arguments{
\item{object}{Signature object}

\item{data}{Additional signature data, in the form of a named list}
}
\value{
Updated Signature object.
}
\description{
Add signature data to the sigDict of this Signature Object
}
