% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{addTSNE}
\alias{addTSNE}
\title{Adds a tSNE projection}
\usage{
addTSNE(object, perplexity = 30, name = "tSNE", source = "LatentSpace")
}
\arguments{
\item{object}{the VISION object}

\item{perplexity}{parameter for tSNE}

\item{name}{label to use for this projection}

\item{source}{coordinates to use to compute tSNE
Choices are either "LatentSpace" (default) or "Proteins"}
}
\value{
VISION object with the projection added to @Projections slot
}
\description{
Adds a tSNE projection
}
