% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{addUMAP}
\alias{addUMAP}
\title{Adds a UMAP projection}
\usage{
addUMAP(object, K = object@params$numNeighbors, name = "UMAP",
  source = "LatentSpace")
}
\arguments{
\item{object}{the VISION object}

\item{K}{Number of neighbors to use in UMAP projection.}

\item{name}{label to use for this projection}

\item{source}{coordinates to use to compute tSNE
Choices are either "LatentSpace" (default) or "Proteins"}
}
\value{
VISION object with the projection added to @Projections slot
}
\description{
Adds a UMAP projection
}
