% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\docType{methods}
\name{Analyze,FastProject-method}
\alias{Analyze,FastProject-method}
\title{Main entry point for running FastProject Analysis}
\usage{
\S4method{Analyze}{FastProject}(object)
}
\arguments{
\item{object}{FastProject object}
}
\value{
FastProjectOutput object
}
\description{
Main entry point for running FastProject Analysis
}
\examples{
fp <- FastProject("expression_matrix.txt", "data/Gene Name Housekeeping.txt", c("sigfile_1.gmt", "sigfile_2.txt"),
                  precomputed="pre_sigs.txt")
fpout <- Analysis(fp)
}
