% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\docType{methods}
\name{Analyze,FastProject-method}
\alias{Analyze,FastProject-method}
\title{Main entry point for running FastProject Analysis}
\usage{
\S4method{Analyze}{FastProject}(object, BPPARAM = NULL)
}
\arguments{
\item{object}{FastProject object}

\item{BPPARAM}{a parallelization backend to use for the analysis}
}
\value{
FastProjectOutput object
}
\description{
The main analysis function. Runs the entire FastProject analysis pipeline
and returns a FastProjectOutput object with the result,
}
\examples{
\dontrun{
fp <- FastProject("data/expression_matrix.txt",
                  "data/Gene Name Housekeeping.txt",
                  c("extdata/geneSetLibrary.gmt",
                    "data/tcga_sigs.txt"))
fpout <- Analysis(fp)
ViewResults(fpout)
}
}
