% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{analyzeHotspotObjectVision}
\alias{analyzeHotspotObjectVision}
\title{Analyze a Hotspot object using built in methods such
such as local correlation, signature overlap, etc.
Necessary to run this function for Hotspot functionality in viewer to work.}
\usage{
analyzeHotspotObjectVision(object, hs, tree = FALSE)
}
\arguments{
\item{object}{the VISION object}

\item{hs}{the Hotspot python object loaded by Reticulate}

\item{tree}{whether to use tree as latent space. If TRUE, object should have a tree}
}
\value{
the modified VISION object with the following slots filled:
Populates the modData, HotspotModuleScores, ModuleSignatureEnrichment
and HotspotObject slots of object, as well as recalculates signature scores
for new modules.
}
\description{
Analyze a Hotspot object using built in methods such
such as local correlation, signature overlap, etc.
Necessary to run this function for Hotspot functionality in viewer to work.
}
