% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{analyzeLocalCorrelations}
\alias{analyzeLocalCorrelations}
\title{Compute local correlations for all signatures}
\usage{
analyzeLocalCorrelations(object, tree = FALSE)
}
\arguments{
\item{object}{the VISION object}
}
\value{
the VISION object with values set for the analysis results
}
\description{
This is the main analysis function. For each filtered dataset, a set of
different projection onto low-dimensional space are computed, and the
consistency of the resulting space with the signature scores is computed
to find signals that are captured succesfully by the projections.
}
