% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ancestor_at_depth}
\alias{ancestor_at_depth}
\title{Find the ancestor of a node above a specific depth}
\usage{
ancestor_at_depth(tree, node, depth)
}
\arguments{
\item{tree}{an object of class phylo}

\item{node}{the index of the node}

\item{depth}{the depth to search to}
}
\value{
the index of the ancestor node
}
\description{
Find the ancestor of a node above a specific depth
}
