% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{annotateLatentComponents}
\alias{annotateLatentComponents}
\title{Compute pearson correlation between signature scores and components of the Latent Space}
\usage{
annotateLatentComponents(object)
}
\arguments{
\item{object}{the VISION object}
}
\value{
object with the @LCAnnotatorData slot populated
}
\description{
Enables the LCAnnotator portion of the output report
}
