% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyICA}
\alias{applyICA}
\title{Performs ICA on data}
\usage{
applyICA(exprData)
}
\arguments{
\item{exprData}{Expression data, NUM_GENES x NUM_SAMPLES}
}
\value{
Reduced data NUM_SAMPLES x NUM_COMPONENTS
}
\description{
Performs ICA on data
}
