% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyISOMap}
\alias{applyISOMap}
\title{Performs ISOMap on data}
\usage{
applyISOMap(exprData, BPPARAM = bpparam())
}
\arguments{
\item{exprData}{Expression data, NUM_GENES x NUM_SAMPLES}

\item{BPPARAM}{the parallelization backend to use}
}
\value{
Reduced data NUM_SAMPLES x NUM_COMPONENTS
}
\description{
Performs ISOMap on data
}
