% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyKNN}
\alias{applyKNN}
\title{create a Knearest neighbor graph from the data}
\usage{
applyKNN(exprData, BPPARAM = bpparam())
}
\arguments{
\item{exprData}{the data to base the KNN graph on}

\item{BPPARAM}{the parallelization backend to use for the computation}
}
\value{
the weghted adjacency matrix of the KNN graph
}
\description{
create a Knearest neighbor graph from the data
}
