% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{applyMicroClustering}
\alias{applyMicroClustering}
\title{Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm}
\usage{
applyMicroClustering(exprData, cellsPerPartition = 100,
  filterInput = "fano", filterThreshold = round(ncol(exprData) * 0.2),
  random = FALSE)
}
\arguments{
\item{exprData}{the expression data matrix}

\item{cellsPerPartition}{control over the minimum number of cells to put into each supercell}
}
\value{
a list:
\itemize{
    \item pooled cells - the super cells creates
    \item pools - a list of data matrices of the original cells in each pool
}
}
\description{
Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm
}
