% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyPCA}
\alias{applyPCA}
\title{Performs PCA on data}
\usage{
applyPCA(exprData, N = 0, variance_proportion = 1)
}
\arguments{
\item{exprData}{Expression data}

\item{N}{Number of components to retain. Default is 0}

\item{variance_proportion}{Retain top X PC's such that this much variance is retained; if N=0, then apply this method}
}
\value{
Matrix containing N components for each sample.
}
\description{
Performs PCA on data
}
