% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyRBFPCA}
\alias{applyRBFPCA}
\title{Performs PCA on data that has been transformed with the Radial Basis Function.}
\usage{
applyRBFPCA(exprData, BPPARAM = bpparam())
}
\arguments{
\item{exprData}{Expression data, NUM_GENES x NUM_SAMPLES}

\item{BPPARAM}{the parallelization backend to use}
}
\value{
Reduced data NUM_SAMPLES x NUM_COMPONENTS
}
\description{
Performs PCA on data that has been transformed with the Radial Basis Function.
}
