% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{applySimplePPT}
\alias{applySimplePPT}
\title{Applies the Simple PPT algorithm onto the expression data.}
\usage{
applySimplePPT(exprData, numCores, permExprData = NULL,
  nNodes_ = round(sqrt(nrow(exprData))), sigma = 0, gamma = 0)
}
\arguments{
\item{exprData}{Expression data -- Num_Genes x Num_Samples}

\item{numCores}{Number of cores to use during this analysis}

\item{permExprData}{a list of permutated expression datasets,
to use for significance estimation of the tree [default:NULL]}

\item{nNodes_}{Number of nodes to find. Default is sqrt(N)}

\item{sigma}{regularization parameter for soft-assignment of data points to
nodes, used as the variance
         of a guassian kernel. If 0, this is estimated automatically}

\item{gamma}{graph-level regularization parameter, controlling the tradeoff
between the noise-levels
         in the data and the graph smoothness. If 0, this is estimated
         automatically.}
}
\value{
Information on the fitten tree
     \itemize{
    \item C: spatial positions of the tree nodes in NUM_PROTEINS dimensional
    space
    \item W: Unweighted (binary) adjacency matrix of the fitten tree
    \item distMat: distance matrix between each tree node to each datapoint
    \item mse: the Mean-Squared-Error of the fitten tree
    \item zscore: a significance score for the fitted tree
     }
}
\description{
Applies the Simple PPT algorithm onto the expression data.
}
