% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applySpectralEmbedding}
\alias{applySpectralEmbedding}
\title{Performs Spectral Embedding  on data}
\usage{
applySpectralEmbedding(exprData)
}
\arguments{
\item{exprData}{Expression data, NUM_GENES x NUM_SAMPLES}
}
\value{
Reduced data NUM_SAMPLES x NUM_COMPONENTS
}
\description{
Performs Spectral Embedding  on data
}
