% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applytSNE10}
\alias{applytSNE10}
\title{Performs tSNE with perplexity 10 on data}
\usage{
applytSNE10(exprData, BPPARAM = bpparam())
}
\arguments{
\item{exprData}{Expression data, NUM_GENES x NUM_SAMPLES}

\item{BPPARAM}{the parallelization backend to use}
}
\value{
Reduced data NUM_SAMPLES x NUM_COMPONENTS
}
\description{
Performs tSNE with perplexity 10 on data
}
