% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyUMAP}
\alias{applyUMAP}
\title{Performs UMAP on data}
\usage{
applyUMAP(exprData)
}
\arguments{
\item{exprData}{Expression data, NUM_GENES x NUM_SAMPLES}
}
\value{
Reduced data NUM_SAMPLES x NUM_COMPONENTS
}
\description{
Performs UMAP on data
}
