% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyWeightedPCA}
\alias{applyWeightedPCA}
\title{Performs weighted PCA on data}
\usage{
applyWeightedPCA(exprData, weights, maxComponents = 200)
}
\arguments{
\item{exprData}{Expression matrix}

\item{weights}{Weights to use for each coordinate in data}

\item{maxComponents}{Maximum number of components to calculate}
}
\value{
Weighted PCA data

Variance of each component

Eigenvectors of weighted covariance matrix, aka the variable loadings
}
\description{
Performs weighted PCA on data
}
