% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{batchify}
\alias{batchify}
\title{Utility methods
Helper utility to group list items into batches}
\usage{
batchify(items, per_batch, n_workers = 1)
}
\arguments{
\item{items}{list of items to group into batches}

\item{per_batch}{approximate target for size of batches}

\item{n_workers}{number of batches is made divisible by n_workers}
}
\value{
list of list of items
}
\description{
This is used inside the batchSigEval function
}
