% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{batchSigEval}
\alias{batchSigEval}
\title{Evaluate signature scores efficiently in batches}
\usage{
batchSigEval(sigs, sig_score_method, eData, weights, min_signature_genes)
}
\arguments{
\item{sigs}{list of Signature(s) to be evalauting}

\item{sig_score_method}{either "naive" or "weighted_avg"}

\item{eData}{numeric Matrix Genes x Cells}

\item{weights}{Weight matrix computed through FNR curve}

\item{min_signature_genes}{signatures are discarded which do not match at least
this many genes in the expression matrix}
}
\value{
List of SignatureScore objects
}
\description{
Evaluate signature scores efficiently in batches
}
