% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{batchSigEvalNorm}
\alias{batchSigEvalNorm}
\title{Evaluate signature scores efficiently in batches}
\usage{
batchSigEvalNorm(sigs, normData)
}
\arguments{
\item{sigs}{list of Signature(s) to be evalauting}

\item{normData}{NormData object}
}
\value{
matrix of signature scores, cells X signatures
}
\description{
This version uses the NormData object to operate
without inflating sparse matrices
}
