% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{biasedSD}
\alias{biasedSD}
\title{Calculates standard deviation with a denominator of n rather than (n-1)}
\usage{
biasedSD(data, byRow = TRUE)
}
\arguments{
\item{data}{data matrix to apply the biased standard deviation operation to}

\item{byRow}{if TRUE, apply by row. Else, by column. Default is TRUE.}
}
\value{
Vector of standard deviations by row or column
}
\description{
Calculates standard deviation with a denominator of n rather than (n-1)
}
