% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{biasedVectorSD}
\alias{biasedVectorSD}
\title{Computes the biased SD on a vector, correcting the denominator to n rather than (n-1)}
\usage{
biasedVectorSD(data)
}
\arguments{
\item{data}{Vector of numbers}
}
\value{
Biased standard deviation of vector
}
\description{
Computes the biased SD on a vector, correcting the denominator to n rather than (n-1)
}
