% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{calc_mod_sig_enrichment}
\alias{calc_mod_sig_enrichment}
\title{Computes the hypergeometric overlap test for modules and signatures}
\usage{
calc_mod_sig_enrichment(object, skip_down = TRUE)
}
\arguments{
\item{object}{the Vision object.}

\item{skip_down}{whether to ignore down signatures in overlap}
}
\value{
list(statistic values, p values, clusters of signatures)
}
\description{
Computes the hypergeometric overlap test for modules and signatures
}
