% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{calc_set_enrichment}
\alias{calc_set_enrichment}
\title{Calculate the hypergeometric enrichment for two sets from a population
Statisic = log (observed overlap / expected overlap)
P value = 1- hypergeometric(observed overlap -1, max(|set1|, |set2|), |genes| - |set1|, min(|set1|, |set2|))}
\usage{
calc_set_enrichment(set1, set2, genes)
}
\arguments{
\item{set1}{the first gene set}

\item{set2}{the second gene set}

\item{genes}{the population}
}
\value{
c(statistic, p value)
}
\description{
Calculate the hypergeometric enrichment for two sets from a population
Statisic = log (observed overlap / expected overlap)
P value = 1- hypergeometric(observed overlap -1, max(|set1|, |set2|), |genes| - |set1|, min(|set1|, |set2|))
}
