% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{calcInterEdgeDistMat}
\alias{calcInterEdgeDistMat}
\title{Calculate all distances between points on two different edges}
\usage{
calcInterEdgeDistMat(v1.dist, v2.dist, path.length)
}
\arguments{
\item{v1.dist}{a vectors of all distances between all points on the first edge,
and the vertex of the first edge that is closer to the second edge}

\item{v2.dist}{a vectors of all distances between all points on the second edge,
and the vertex of the second edge that is closer to the first edge}
}
\description{
Calculate all distances between points on two different edges
}
