% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{calcIntraEdgeDistMat}
\alias{calcIntraEdgeDistMat}
\title{Calculate distances between all points on a given edge, including edge vertices}
\usage{
calcIntraEdgeDistMat(edge.len, edgePos)
}
\arguments{
\item{edge.len}{the length of the node}

\item{edgePos}{the relative positions of the points on the edge (in range [0,1]).}
}
\value{
a distance matrix, where the first and last points are the edge vertices
}
\description{
Calculate distances between all points on a given edge, including edge vertices
}
