% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{calcPearsonCorrelation}
\alias{calcPearsonCorrelation}
\title{Compute pearson correlation between signature scores and principle components}
\usage{
calcPearsonCorrelation(ss, pc)
}
\arguments{
\item{ss}{signature scores}

\item{pc}{principle components}
}
\value{
pearson correlaton coeffcients
}
\description{
Compute pearson correlation between signature scores and principle components
}
