% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{calculatePearsonCorr}
\alias{calculatePearsonCorr}
\title{Compute pearson correlation between signature scores and principle components}
\usage{
calculatePearsonCorr(sigData, sigMatrix, fullPCA)
}
\arguments{
\item{sigData}{List of Signature}

\item{sigMatrix}{Signature scores dataframe}

\item{fullPCA}{numeric matrix N_Cells x N_PCs}
}
\value{
pearsonCorr numeric matrix N_Signatures x N_PCs
}
\description{
Compute pearson correlation between signature scores and principle components
}
