% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{calculateSignatureBackground}
\alias{calculateSignatureBackground}
\title{calculate background signature scores}
\usage{
calculateSignatureBackground(object, num, BPPARAM = NULL)
}
\arguments{
\item{object}{the FastProject object}

\item{num}{the number of signatures to generate}

\item{BPPARAM}{the parallelization backend to use for these computations}
}
\value{
numeric matrix of random signature scores SIGNATURES x CELLS
}
\description{
For each signature-cell pair, compute a score that captures the level of
correspondence between the cell and the signature.
To estimate significance of these scores, a set of random gene signatures is
generated to create a null distribution
}
