% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{calculateTrajectoryDistances}
\alias{calculateTrajectoryDistances}
\title{Calculate distance matrix between all pairs of ponts based on their projection onto the tree}
\usage{
calculateTrajectoryDistances(adjMat, edgeAssoc, edgePos, latentPnts = NULL)
}
\arguments{
\item{adjMat}{(K x K logical) adjacency matrix of the milestones}

\item{edgeAssoc}{(2 x N) for each point, the edge it is projected to (represented as (V1,V2), where V1<V2)}

\item{edgePos}{(length N, numeric) relative postion on the edge for each point, in range [0,1]}

\item{latentPnts}{(D x K numeric) the spatial locations of the milestones}
}
\value{
non-negative symmetric matrix in which [i,j] is the tree-based distance between points i, j.
}
\description{
Calculate distance matrix between all pairs of ponts based on their projection onto the tree
}
