% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{calculateTreeDistances}
\alias{calculateTreeDistances}
\title{Calculate distance matrix between all pairs of ponts based on their projection onto the tree}
\usage{
calculateTreeDistances(princPnts, princAdj, edgeAssoc, edgePos)
}
\arguments{
\item{princPnts}{(D x K numeric) the spatial locations of the principle points (tree vertices)}

\item{princAdj}{(K x K logical) adjacency matrix of the principle graph}

\item{edgeAssoc}{(2 x N) for each point, the edge it is projected to (represented as (V1,V2), where V1<V2)}

\item{edgePos}{(length N, numeric) relative postion on the edge for each point, in range [0,1]}
}
\value{
non-negative symmetric matrix in which [i,j] is the tree-based distance between points i, j.
}
\description{
Calculate distance matrix between all pairs of ponts based on their projection onto the tree
}
\examples{
X <- matrix(rnorm(200), nrow = 2)
tree <- applySimplePPT(X)
proj <- projectOnTree(X, tree[[1]], tree[[2]])
distmat <- calculateTreeDistances(tree[[1]], tree[[2]], proj$edges, proj$edgePos)
}
