% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{calcWeights}
\alias{calcWeights}
\title{Calculate weights based on the false negative rates, computed using the
provided housekeeping genes}
\usage{
calcWeights(object)
}
\arguments{
\item{object}{the FastProject object}
}
\value{
the FastProject object with populated weights slot
}
\description{
Calculate weights based on the false negative rates, computed using the
provided housekeeping genes
}
