% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{clipBottom}
\alias{clipBottom}
\title{Sets all values below a certain level in the data equal to 0}
\usage{
clipBottom(x, mi)
}
\arguments{
\item{x}{Data matrix}

\item{mi}{Minimum value}
}
\value{
Data matrix with all values less than MI set to 0
}
\description{
Sets all values below a certain level in the data equal to 0
}
