% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Projection.R
\docType{methods}
\name{cluster,Projection-method}
\alias{cluster,Projection-method}
\title{Clusters the projection according to some method}
\usage{
\S4method{cluster}{Projection}(object, method, param)
}
\arguments{
\item{object}{Projection object}

\item{method}{Method by which to cluster the data}

\item{param}{Parameters for clustering method}
}
\description{
Clusters the projection according to some method
}
