% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{clusterCells}
\alias{clusterCells}
\title{Creates clustering of the cells}
\usage{
clusterCells(object, tree = FALSE)
}
\arguments{
\item{object}{the VISION object for which to cluster the cells}
}
\value{
the VISION object modifed as described above
}
\description{
Uses the Louvain clustering algorithm to cluster the cells in the
transcriptional latent space (default is the top PCs, unless a user
specifies otherwise). If a tree exists, a tree-based clustering is also
performed.
}
\details{
Results of this are stored as a new variable in the object's metaData
}
