% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{clusterSignatures}
\alias{clusterSignatures}
\title{Clusters signatures according to the rank sum}
\usage{
clusterSignatures(sigList, sigMatrix, pvals)
}
\arguments{
\item{sigList}{List of signatures}

\item{sigMatrix}{data.frame of signatures scores, NUM_SAMPLES x NUM_SIGNATURES}

\item{pvals}{the corresponding P-values for each score,
NUM_SIGNATURES x NUM_SAMPLES}
}
\value{
a list:
\itemize{
    \item Computed: a list of clusters of computed signatures
    \item Precomputed: a list of clusters of precomputed signatures
}
}
\description{
Clusters signatures according to the rank sum
}
