% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{clusterSigScores}
\alias{clusterSigScores}
\title{Compute Ranksums Test, for all factor meta data.  One level vs all others}
\usage{
clusterSigScores(object)
}
\arguments{
\item{object}{the VISION object}
}
\value{
the VISION object with values set for the analysis results
}
\description{
Compute Ranksums Test, for all factor meta data.  One level vs all others
}
