% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{clusterToJSON}
\alias{clusterToJSON}
\title{Convert a Cluster object to JSON}
\usage{
clusterToJSON(cluster)
}
\arguments{
\item{cluster}{Cluster object}
}
\value{
Cluster object converted to JSON
}
\description{
Convert a Cluster object to JSON
}
