% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{colNormalization}
\alias{colNormalization}
\title{Performs z-normalization on all columns}
\usage{
colNormalization(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
Data matrix with same dimensions, with each column z-normalized.
}
\description{
Performs z-normalization on all columns
}
