% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{colRankNormalization}
\alias{colRankNormalization}
\title{Creaes a new version of the data that has ranks (column-wise) instead of values.}
\usage{
colRankNormalization(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
Data matrix with same dimensions, with each value representing its column-wise rank.
}
\description{
Creaes a new version of the data that has ranks (column-wise) instead of values.
}
