% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{colVarsSp}
\alias{colVarsSp}
\title{Compute col-wise variance on matrix without densifying}
\usage{
colVarsSp(x)
}
\arguments{
\item{x}{expression matrix}
}
\value{
numeric vector col-wise variance
}
\description{
Compute col-wise variance on matrix without densifying
}
