% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\docType{methods}
\name{computeKNNWeights,matrix-method}
\alias{computeKNNWeights,matrix-method}
\title{compute for each vector the weights to apply to it's K nearest neighbors}
\usage{
\S4method{computeKNNWeights}{matrix}(object, K = round(sqrt(nrow(object))))
}
\arguments{
\item{object}{matrix to use for KNN}

\item{K}{number of neighbors to compute this for}
}
\value{
a weights matrix
}
\description{
compute for each vector the weights to apply to it's K nearest neighbors
}
