% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{computeKNNWeights,phylo-method}
\alias{computeKNNWeights,phylo-method}
\title{Compute for each vector the weights to apply to it's K nearest neighbors}
\usage{
\S4method{computeKNNWeights}{phylo}(
  object,
  K = round(sqrt(length(object$tip.label))),
  lcaKNN = FALSE,
  minSize = 20
)
}
\arguments{
\item{object}{tree to use for KNN}

\item{K}{Number of neighbors to consider.}

\item{lcaKNN}{whether to use LCA based KNN (cluster by minimum size), if false defaults to cophenetic distance (random tie breaking).
WARNING: lcaKNN doesn't perform well with broad multifurcations}
}
\value{
a list of two items:
         indices: matrix, cells X neighbors
             Each row specifies indices of nearest neighbors
         weights: matrix, cells X neighbors
             Corresponding weights to nearest neighbors
}
\description{
Compute for each vector the weights to apply to it's K nearest neighbors
}
