% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Projection.R
\docType{methods}
\name{computeKNNWeights,Projection-method}
\alias{computeKNNWeights,Projection-method}
\title{compute for each vector the weights to apply to it's K nearest neighbors}
\usage{
\S4method{computeKNNWeights}{Projection}(object, K = 30)
}
\arguments{
\item{object}{the Projecton object}

\item{K}{number of neughbors to compute ths for}
}
\value{
a weights matrix
}
\description{
compute for each vector the weights to apply to it's K nearest neighbors
}
