% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TreeProjection.R
\docType{methods}
\name{computeKNNWeights,TreeProjection-method}
\alias{computeKNNWeights,TreeProjection-method}
\title{Compute KNN weights based on geodesic distances for TreeProjection objects}
\usage{
\S4method{computeKNNWeights}{TreeProjection}(object,
  K = round(sqrt(nrow(object@pData))))
}
\arguments{
\item{object}{a TreeProjection object}

\item{K}{the number of nearest neighbors to look at}
}
\value{
a list of two items:
         indices: matrix, cells X neighbors
             Each row specifies indices of nearest neighbors
         weights: matrix, cells X neighbors
             Corresponding weights to nearest neighbors
}
\description{
Compute KNN weights based on geodesic distances for TreeProjection objects
}
