% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{computeLatentSpace}
\alias{computeLatentSpace}
\title{Computes the latent space of the expression matrix using PCA}
\usage{
computeLatentSpace(object, projection_genes = NULL, perm_wPCA = NULL)
}
\arguments{
\item{object}{the VISION object for which compute the latent space}

\item{projection_genes}{character vector of gene names to use for projections}

\item{perm_wPCA}{If TRUE, apply permutation wPCA to determine significant
number of components. Default is FALSE.}
}
\value{
the VISION with latentSpace populated
}
\description{
Computes the latent space of the expression matrix using PCA
}
