% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{computeProjectionGenes}
\alias{computeProjectionGenes}
\title{filter data accourding to the provided filters}
\usage{
computeProjectionGenes(object, threshold = NULL, num_mad = NULL,
  projection_genes_method = NULL, projection_genes = NULL)
}
\arguments{
\item{object}{the VISION object}

\item{threshold}{Threshold to apply when using the 'threshold' or 'fano' projection genes filter.
If greater than 1, this specifies the number of cells in which a gene must be detected
for it to be used when computing PCA. If less than 1, this instead specifies the proportion of cells needed}

\item{num_mad}{Number of median absolute deviations to use when selecting highly-variable
genes in each mean-sorted bin of genes}

\item{projection_genes_method}{a method to select genes either 'Threshold' or 'Fano'}

\item{projection_genes}{a list of genes to use specifically.
If this is supplied then `projection_genes_method` is ignored}
}
\value{
the VISION object, populated with filtered data
}
\description{
filter data accourding to the provided filters
}
