% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{coordinatesToJSON}
\alias{coordinatesToJSON}
\title{Converts a projection into a JSON object mapping each sample to a projection coordinate.}
\usage{
coordinatesToJSON(p)
}
\arguments{
\item{p:}{Projection coordinate data (NUM_SAMPLES x NUM_COMPONENTS)}
}
\value{
JSON object mapping each sample to a projection coordinate.
}
\description{
Converts a projection into a JSON object mapping each sample to a projection coordinate.
}
