% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{createGeneSignature}
\alias{createGeneSignature}
\title{Create a user-defined gene signature}
\usage{
createGeneSignature(name, sigData, metadata = "")
}
\arguments{
\item{name}{the name of the signature}

\item{sigData}{a named vector where the names correspond to genes in the
data and the values are either `1.0` for up-regulated (or positive) genes,
and `-1.0` for down-regulated (negtive) genes. For an unsigned signature
use 1.0 for all values.}

\item{metadata}{metadata that is relevent to the signature. [Default:NULL]}
}
\value{
a Signature object
}
\description{
Typical usage of VISION involes providing a location to a ".gmt" signature
file from which VISION automatically creates Signature objects.  However,
using the createGeneSignature method, users may programmatically define
signatures from within R.
}
\examples{


sig1 <- createGeneSignature(
           name = "CD8 Markers",
           sigData = c(CD8A=1, CD8B=1, GZMK=1, GZMB=1,
                       GZMH=1, GZMA=1, GNLY=1, DUSP2=1,
                       EOMES=1, TBX21=1, PRMD1=1, PRF1=1,
                       IFNG=1)
       )

cc_sigs <- "path/to/cell_cycle.gmt"

sigs <- c(sig1, cc_sigs)

vis <- Vision(data = expMat, signatures = sigs)


}
