% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{createNewFP}
\alias{createNewFP}
\title{Creates a new FastProject object with the new expression matrix. Called from
Server.R.}
\usage{
createNewFP(fpParams, nexpr)
}
\arguments{
\item{fpParams}{List of FastProject parameters}

\item{nexpr}{New expression matrix for the new FastProject object}
}
\value{
a new FastProject object
}
\description{
Creates a new FastProject object with the new expression matrix. Called from
Server.R.
}
