% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\docType{methods}
\name{createOutputDirectory,FastProject-method}
\alias{createOutputDirectory,FastProject-method}
\title{create an output directory in the working directory}
\usage{
\S4method{createOutputDirectory}{FastProject}(object)
}
\arguments{
\item{object}{the FastProject object the requires an output directory
Creates the output directory structure}
}
\description{
create an output directory in the working directory
}
