% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{createPooledMetaData}
\alias{createPooledMetaData}
\title{Aggregate meta-data for cells in pools}
\usage{
createPooledMetaData(metaData, pools, cluster_variable = "")
}
\arguments{
\item{metaData}{data.frame of meta-data for cells}

\item{pools}{named list of character vector describing cell ids
in each pool}

\item{cluster_variable}{name of metaData variable that was used
to pre-parition cells before micropooling}
}
\value{
a data.frame of new meta-data
}
\description{
Aggregate meta-data for cells in pools
}
