% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{createPoolsBatch}
\alias{createPoolsBatch}
\title{create "super-cells" by pooling together single cells}
\usage{
createPoolsBatch(cl, expr)
}
\arguments{
\item{cl}{cluster association of each cell}

\item{expr}{expression data}
}
\value{
a matrx of expression data for the pooled cells
}
\description{
This function calls createPools on batches of clusters in parallel
}
