% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Trajectory.R
\name{createTrajectoryMetaData}
\alias{createTrajectoryMetaData}
\title{Generate meta-data associated with this trajectory}
\usage{
createTrajectoryMetaData(trajectory)
}
\arguments{
\item{trajectory}{Trajectory on which to operate}
}
\value{
metaData dataframe with meta-data
}
\description{
Creates a categorical variable mapping cells to edges
and numeric variables for their position along edges
}
