% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{createUserGeneSignature}
\alias{createUserGeneSignature}
\title{Create a user-defined gene signature}
\usage{
createUserGeneSignature(name, sigData, metadata = "")
}
\arguments{
\item{name}{the name of the signature}

\item{sigData}{a named vector where the names correspond to genes in the
data and the values are either `1.0` for up-regulated (or positive) genes,
and `-1.0` for down regulated (negtive) genes.}

\item{metadata}{metadata that is relevent to the signature. [Default:NULL]}
}
\value{
a Signature object
}
\description{
Create a user-defined gene signature
}
\examples{
sig <- createUserGeneSignature(name="signature", c(gene1=1.0,gene17=1.0,
                                                   gene4=1.0, gene31=-1.0,
                                                   gene3=-1.0, gene9=1.0))
sig2 <- createUserGeneSignature(name="signature", c(gene18=1.0,gene29=1.0,
                                                   gene400=-1.0, gene1=-1.0,
                                                   gene7=1.0, gene9=1.0))
}
