% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{depthBasedCladewiseTreeCluster}
\alias{depthBasedCladewiseTreeCluster}
\title{Performs a breadth first search to create a specific number of clusters.
Clusters are split based on depth.}
\usage{
depthBasedCladewiseTreeCluster(tree, target = 10)
}
\arguments{
\item{tree}{object of class phylo}

\item{target}{number of clusters to attempt to generate}
}
\value{
List of clusters, each entry being a vector of indices representing
samples in the cluster.
}
\description{
Performs a breadth first search to create a specific number of clusters.
Clusters are split based on depth.
}
