% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{depthBasedTreeCluster}
\alias{depthBasedTreeCluster}
\title{Performs a binary search on a depth d such that 
if depth(LCA(u, v)) <= d then u and v are in the same cluster}
\usage{
depthBasedTreeCluster(tree, target = 10)
}
\arguments{
\item{tree}{object of class phylo}

\item{target}{number of clusters to attempt to generate}
}
\value{
List of clusters, each entry being a vector of indices representing
samples in the cluster.
}
\description{
Performs a binary search on a depth d such that 
if depth(LCA(u, v)) <= d then u and v are in the same cluster
}
