% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{diffusionTime}
\alias{diffusionTime}
\title{Computes diffusion pseudotime.}
\usage{
diffusionTime(counts, root, k = NULL)
}
\arguments{
\item{counts}{TreeProjection matrix}

\item{k}{# of neighbors for kernel width. Default sqrt(n).}
}
\value{
list:
\itemize{
    \item path i: a list containing the cells and pseudotime coordinates
    for each path originating from root.
}
}
\description{
Computes diffusion pseudotime.
}
