% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{.colNormHelper}
\alias{.colNormHelper}
\title{Calculates the column znormalization after row znormalization}
\usage{
.colNormHelper(data, rowOffsets, rowScaleFactors)
}
\arguments{
\item{data}{expression data matrix}

\item{rowOffsets}{offsets to be subtracted from each row}

\item{rowScaleFactors}{factors to scale each row}
}
\value{
colOffsets and colScaleFactors
}
\description{
Helper to do this without inflating the matrix
}
