% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{draw_hotspot_heatmap}
\alias{draw_hotspot_heatmap}
\title{Draw Modules Heatmap (Gene x Gene)}
\usage{
draw_hotspot_heatmap(hs, palette = paletteer_d("ggsci::default_nejm"))
}
\arguments{
\item{hs}{the Hotspot Object}

\item{palette}{palette}
}
\description{
Draw Modules Heatmap (Gene x Gene)
}
