% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{evalSigGeneImportance}
\alias{evalSigGeneImportance}
\title{calculate gene-signature importance}
\usage{
evalSigGeneImportance(object)
}
\arguments{
\item{object}{the VISION object}
}
\value{
the VISION object, with SigGeneImportance slot populated
}
\description{
For each signature, the contribution of each gene to the signature score
is evaluated by calculating the covariance between signature scores and expression
The correlation of genes with a negative sign in the signature are inverted.
}
