% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{evalSigGeneImportanceSparse}
\alias{evalSigGeneImportanceSparse}
\title{Calculate Gene-Signature Importance}
\usage{
evalSigGeneImportanceSparse(sigScores, sigData, normExpr)
}
\arguments{
\item{sigScores}{matrix of signature scores (Cells x Signatures)}

\item{sigData}{list of Signature objects}

\item{normExpr}{output from calling getNormalizedCopySparse}

\item{object}{the VISION object}
}
\value{
the VISION object, with SigGeneImportance slot populated

named list with one item per signature.  Values are named vectors
of each gene's association (covariance) with the signature.
}
\description{
For each signature, the contribution of each gene to the signature score
is evaluated by calculating the covariance between signature scores and expression
The correlation of genes with a negative sign in the signature are inverted.
}
\details{
This version is made to avoid inflating sparse matrices
}
