% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\name{ExpressionData}
\alias{ExpressionData}
\title{Initializes an ExpressionData object
should not be called directly, use the `new` syntax}
\usage{
ExpressionData(data, ...)
}
\arguments{
\item{data}{expression data matrix}

\item{...}{additional arguments}
}
\value{
ExpressionData object
}
\description{
Initializes an ExpressionData object
should not be called directly, use the `new` syntax
}
