% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{expressionToJSON}
\alias{expressionToJSON}
\title{Convertes expression matrix to JSON}
\usage{
expressionToJSON(expr, geneList = NULL, zscore = FALSE)
}
\arguments{
\item{expr}{Expression Matrix}

\item{geneList}{optional list of genes to subset from expr}
}
\value{
(Potentially) subsetted expression matrix
}
\description{
Convertes expression matrix to JSON
}
