% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FastProjectOutput-class}
\alias{FastProjectOutput-class}
\alias{FastProjectOutput}
\title{a container for FastProject output data}
\description{
a container for FastProject output data
}
