% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-FastProjectOutput.R
\name{FastProjectOutput}
\alias{FastProjectOutput}
\alias{FastProjectOutput}
\title{a container for FastProject output data}
\usage{
FastProjectOutput(eData, filterModuleList, sigMatrix, sigList, fpParams, pools)

FastProjectOutput(eData, filterModuleList, sigMatrix, sigList, fpParams, pools)
}
\arguments{
\item{eData}{ExpressionData matrix}

\item{filterModuleList}{List of FilterModuleData objects}

\item{sigMatrix}{Matrix of signature scores (NUM_SIGS x NUM_CELLS)}

\item{sigList}{List of Signatures}

\item{fpParams}{List of the FastProject parameters that were used to
generate this object}

\item{pools}{a list of pooling information for pooling single cells into
representative super-cell clusters}
}
\value{
FastProjectOutput object.
}
\description{
a container for FastProject output data

Initialize a new FastProjectOuptput object
should not be called directly, use the `new` syntax
}
