% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{FastProjectR}
\alias{FastProjectR}
\alias{FastProjectR-package}
\title{FastProjectR}
\description{
FastProjectR analyzes expression data and produces a dynamic report in which various two dimensional
projections of the data can be explored. Annotated gene sets ("signatures") are incorporated so projection features
can be understood in relation to the biological processes present in the data. In addition to standard dimensionality
reduction methods, FastProjectR incorporates fitting a high dimensional principle tree to allow exploring continuous
processes. FastProjectR supports the analysis of massive datasets - theoretically up to a million single cells at any
sequencing depth, and an interactive visualization platforms that enables subset analysis based on two dimensional
representations.
}
