% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{fbConsistencyScores}
\alias{fbConsistencyScores}
\title{Evaluates the significance of each feature barcode}
\usage{
fbConsistencyScores(weights, featureBarcodeData)
}
\arguments{
\item{weights}{output of computeKNNWeights}

\item{featureBarcodeData}{numeric matrix of feature barcode counts
size is cells x features}
}
\value{
dataframe with columns "C", "pValue", and "FDR"
}
\description{
Evaluates the significance of each feature barcode
}
