% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{filterData}
\alias{filterData}
\title{filter data accourding to the provided filters}
\usage{
filterData(object, threshold = object@threshold, filters = object@filters)
}
\arguments{
\item{object}{the FastProject object}

\item{threshold}{threshold to apply for the threshold filter}

\item{filters}{list of filters to compute options are limited to}
}
\value{
the FastProject object, populated with filtered data
}
\description{
filter data accourding to the provided filters
}
