% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterGenesNovar}
\alias{filterGenesNovar}
\title{Eliminate genes whose sample variance is equal to 0 (may remove rows);
run when --nofilter option
is selected}
\usage{
filterGenesNovar(data)
}
\arguments{
\item{data}{expression matrix}
}
\value{
character vector of gene names passing filter
}
\description{
Eliminate genes whose sample variance is equal to 0 (may remove rows);
run when --nofilter option
is selected
}
