% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{find_knn_parallel_tree}
\alias{find_knn_parallel_tree}
\title{Parallel KNN for Trees}
\usage{
find_knn_parallel_tree(tree, K)
}
\arguments{
\item{tree}{an object of class phylo}

\item{K}{number of neighbors to query}
}
\value{
list with two items:
    index: Samples x K matrix of neighbor indices
    dist: Samples x K matrix of neighbor distances
}
\description{
Computes nearest-neighbor indices and distances
in parallel.  Query is over all points and results
do not include the self-distance.
}
