% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{find_root}
\alias{find_root}
\title{Find the root node}
\usage{
find_root(tree)
}
\arguments{
\item{tree}{an object of class phylo}
}
\value{
the index of the root node
}
\description{
Find the root node
}
