% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{findNeighbors}
\alias{findNeighbors}
\title{Find K nearest neighbors for each vector in query among the vectors in the
given data}
\usage{
findNeighbors(data, query, k)
}
\arguments{
\item{data}{the search-space to look for nearest neighbors}

\item{query}{the points to find neighbors for (not necessarily in the data)}

\item{k}{the number of neighbors to find for each vectors}
}
\value{
a list of the k neughbors for each of the vectors in `query`
}
\description{
Find K nearest neighbors for each vector in query among the vectors in the
given data
}
