% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{findRepSubset}
\alias{findRepSubset}
\title{find a representative subset of numbers from a given vector}
\usage{
findRepSubset(ls)
}
\arguments{
\item{ls}{a lst of numbers to find a minimal representation of their distibution}
}
\value{
a list of representatve numbers
}
\description{
find a representative subset of numbers from a given vector
}
