% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{fitTree}
\alias{fitTree}
\title{Fit tree using input parameters}
\usage{
fitTree(expr, nNodes, sigma, gamma, tol, maxIter)
}
\arguments{
\item{expr}{Data to fit (NUM_GENES x NUM_SAMPLES)}

\item{nNodes}{Number of nodes in the fitted tree, default is square-root of number of data points}

\item{sigma}{Regularization parameter for soft-assignment of data points to nodes, used as the
variance of a gaussian kernel. If 0, this is estimated automatically.}

\item{gamma}{Graph-level regularization parameter, controlling the tradeoff between the noise-levels
in the data and the graph smoothness. If 0, this is estimated automatically}

\item{tol}{Tolerance to use when fitting the tree}

\item{maxIter}{Maximum number of Iterations ot run the algorithm for}
}
\value{
(list) Tree characteristics:
     \itemize{
    \item C spatial positions of the tree nodes in NUM_PROTEINS dimensional space
    \item unweighted (binary) adjacency matrix
    \item the mean-squared error of the tree
     }
}
\description{
Fit tree using input parameters
}
