% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{geary_sig_v_proj}
\alias{geary_sig_v_proj}
\title{Evaluates values vs coordinates using the Geary C}
\usage{
geary_sig_v_proj(values, indices, weights)
}
\arguments{
\item{values}{numeric matrix of dimension N_SAMPLES x N_SIGNATURES}

\item{indices}{numeric matrix of dimension N_SAMPLES x N_NEIGHBORS}

\item{weights}{numeric matrix of dimension N_SAMPLES x N_NEIGHBORS}
}
\value{
gearyC test statistic values for each signature.  Numeric vector
of length N_SIGNATURES
}
\description{
Evaluates values vs coordinates using the Geary C
}
