% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{generateOverlapSignatures}
\alias{generateOverlapSignatures}
\title{Generates signature objects for the overlap sets between modules and signatures}
\usage{
generateOverlapSignatures(object)
}
\arguments{
\item{object}{the Vision object}
}
\value{
Vision Object, populates the modData slot with overlap signatures.
}
\description{
Generates signature objects for the overlap sets between modules and signatures
}
