% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{generateProjections}
\alias{generateProjections}
\title{Projects data into 2 dimensions using a variety of linear and non-linear methods.}
\usage{
generateProjections(expr, weights, latentSpace, projection_genes = NULL,
  inputProjections = c(), lean = FALSE)
}
\arguments{
\item{expr}{numeric matrix of gene expression}

\item{weights}{weights estimated from FNR curve}

\item{latentSpace}{numeric matrix cells x components}

\item{projection_genes}{character vector of gene names to use for projections}

\item{inputProjections}{Precomputed projections}

\item{lean}{If TRUE, diminished number of algorithms applied,
if FALSE all algorithms applied. Default is FALSE}
}
\value{
list of Projection objects
}
\description{
Projects data into 2 dimensions using a variety of linear and non-linear methods.
}
