% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Trajectory.R
\name{generateTrajectoryProjections}
\alias{generateTrajectoryProjections}
\title{Generate 2d representations of a trajectory model}
\usage{
generateTrajectoryProjections(trajectory)
}
\arguments{
\item{trajectory}{Trajectory on which to operate}
}
\value{
trajectoryProjections list of TrajectoryProjection
}
\description{
Creates 2d layouts of the milestone network and translates the
cell positions into these layouts
}
