% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{generateTreeProjections}
\alias{generateTreeProjections}
\title{Genrate projections based on a tree structure learned in high dimensonal space}
\usage{
generateTreeProjections(data, inputProjections, permMats = NULL)
}
\arguments{
\item{data}{the data to fit the tree to. Should be lower dimensional than
full data.}

\item{inputProjections}{a list of matrices, cells X 2
corresponding TreeProjection will be created. The scores are based
on geodesic distances instead of euclidean one}

\item{permMats}{a list of matrices to use as a null distribution for
estimating the significance of the fitted tree. These are generated by the
permutation wPCA algorithm upstream}
}
\value{
a list:
\itemize{
    \item latentTree TreeProjection object for the main tree
    \item projections a list of TreeProjection objects, mapping the latent
          into the projection's 2d space for visualization purposes
    \item treeScore a score representing the singificance of the fitten tree
    return(list(projections = output, treeScore = hdTree$zscore))
}
}
\description{
Genrate projections based on a tree structure learned in high dimensonal space
}
