% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateBranchDataSet.R
\name{genRandomSigs}
\alias{genRandomSigs}
\title{Utility function to generate random sigs. Allows use of FastProject projections
with unknow signatures}
\usage{
genRandomSigs(expr, nGroups, nGenesPerSig)
}
\arguments{
\item{expr}{matrix numeric genes X cells}

\item{nGroups}{numeric the number of groups to generate sigs}

\item{nGenesPerSig}{numeric the number of genes per signature}
}
\value{
list:
\itemize{
    \item path i: a list containing the cells and pseudotime coordinates
    for each path originating from root.
}
}
\description{
Utility function to generate random sigs. Allows use of FastProject projections
with unknow signatures
}
