% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_all_children}
\alias{get_all_children}
\title{Get all the tip children of a node.}
\usage{
get_all_children(tree, node)
}
\arguments{
\item{tree}{an object of class phylo}

\item{node}{the node to check}
}
\value{
the tips
}
\description{
Get all the tip children of a node.
}
