% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_children}
\alias{get_children}
\title{Find the children of a node}
\usage{
get_children(tree, node)
}
\arguments{
\item{tree}{an object of class phylo}

\item{node}{the node to get the children of}
}
\value{
the children
}
\description{
Find the children of a node
}
