% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_min_cluster_size}
\alias{get_min_cluster_size}
\title{Tree method for getting the min child clade size of a node}
\usage{
get_min_cluster_size(tree, node)
}
\arguments{
\item{tree}{an object of class phylo}

\item{node}{the node to check}
}
\value{
the size of the smallest child clade
}
\description{
Tree method for getting the min child clade size of a node
}
