% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_parent}
\alias{get_parent}
\title{Get the parent of a node}
\usage{
get_parent(tree, node)
}
\arguments{
\item{tree}{an object of class phylo}

\item{node}{the node to get parent}
}
\value{
the immediate parent of the mode, or the node if it is root
}
\description{
Get the parent of a node
}
