% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{getBGDist}
\alias{getBGDist}
\title{Generates, or if already generated retrieves, a random Background Distribution}
\usage{
getBGDist(N_SAMPLES, NUM_REPLICATES)
}
\arguments{
\item{N_SAMPLES}{Number of samples to generate the distribution for}

\item{NUM_REPLICATES}{Number of replicates to generate for the background distribution}
}
\value{
Random matrix with dimensions N_SAMPLES x NUM_REPLICATES with row ordered in ascending order
}
\description{
Generates, or if already generated retrieves, a random Background Distribution
}
\examples{
bgdist <- getBGDist(430, 1000)
}
