% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{getCellVert}
\alias{getCellVert}
\title{Utility function for finding the vertex for a given cell in a projection}
\usage{
getCellVert(projection, cell)
}
\arguments{
\item{projection}{TreeProjection object}

\item{cell}{character name of cell}
}
\value{
numeric: closest vertex for given cell in projection
}
\description{
Utility function for finding the vertex for a given cell in a projection
}
