% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\docType{methods}
\name{getExprData,ExpressionData-method}
\alias{getExprData,ExpressionData-method}
\title{Extracts the expression data from the ExpressionData object}
\usage{
\S4method{getExprData}{ExpressionData}(object)
}
\arguments{
\item{object}{ExpressionData object}
}
\value{
Expression data matrix
}
\description{
Extracts the expression data from the ExpressionData object
}
