% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{getLatentSpace,Vision-method}
\alias{getLatentSpace,Vision-method}
\alias{getLatentSpace}
\title{Get Latent Space}
\usage{
\S4method{getLatentSpace}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
the latent space as a matrix of dimension (Cells x Components)
}
\description{
Provides access to the latent space used for
local autocorrelation analysis
}
\details{
If a latent trajectory was supplied, access it by using \code{getLatentTrajectory}
instead
}
