% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{getLatentTrajectory,Vision-method}
\alias{getLatentTrajectory,Vision-method}
\alias{getLatentTrajectory}
\title{Get Latent Trajectory}
\usage{
\S4method{getLatentTrajectory}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
Trajectory object
}
\description{
Provides access to the latent trajectory used for
local autocorrelation analysis
}
\details{
If a latent space was supplied, access it by using \code{getLatentSpace}
instead
}
\examples{
\dontrun{

trajectory <- getLatentTrajectory(vis)

# MxM connectivity for network milestones
trajectory@adjMat

# data.frame with the position of cells between milestones
# Columns are:
#    cell
#    from (milestone)
#    to (milestone)
#    position (0 to 1)
trajectory@progressions

}
}
