% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\name{getMetaAutocorrelation,Vision-method}
\alias{getMetaAutocorrelation,Vision-method}
\alias{getMetaAutocorrelation}
\title{Get MetaData Autocorrelation Scores}
\usage{
\S4method{getMetaAutocorrelation}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
data.frame with columns 'C', 'pValue', and 'FDR'
}
\description{
Access the local autocorrelation scores computed for meta-data variables
}
\details{
Local autocorrelation scores are calculated from the input latent
space (default's to PCA) or the input trajectory model (if provided)
}
