% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{getMetaConsistency,Vision-method}
\alias{getMetaConsistency,Vision-method}
\alias{getMetaConsistency}
\title{Get MetaData Consistency Scores}
\usage{
\S4method{getMetaConsistency}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
data.frame with columns 'Consistency', 'pValue', and 'FDR'
}
\description{
Access the local consistency scores computed for meta-data variables
}
\details{
Local consistency scores are calculated from the input latent
space (default's to PCA) or the input trajectory model (if provided)
}
