% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\name{getMetaDifferential,Vision-method}
\alias{getMetaDifferential,Vision-method}
\alias{getMetaDifferential}
\title{Get Results of One-vs-All Differential Tests with Metadata Variables}
\usage{
\S4method{getMetaDifferential}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
nested list of list of data.frame (see details)
}
\description{
Returns the results of running one-vs-all differential
tests for each level of every factor meta-variable.
}
\details{
For numeric meta-variables, the 'stat' is the AUC. For factor meta-variables
the stat is the chisq statistic comparing the two groups

The output object has a nested structure:

List of meta-data variables -> List of variable levels -> Results Dataframe

The results dataframe has three columns: "stat", "pValue", "FDR"
}
