% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{getMSE}
\alias{getMSE}
\title{Calculates the MSE between C and X}
\usage{
getMSE(C, X)
}
\arguments{
\item{C}{d x m matrix}

\item{X}{d x n matrix}
}
\value{
Mean Squared Error between C and X.
}
\description{
Calculates the MSE between C and X
}
