% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\docType{methods}
\name{getNormalizedCopy,ExpressionData-method}
\alias{getNormalizedCopy,ExpressionData-method}
\title{Calculates the specified normalized data matrix}
\usage{
\S4method{getNormalizedCopy}{ExpressionData}(object, func)
}
\arguments{
\item{object}{ExpressionData object}

\item{func}{normalization method to apply}
}
\value{
Normalized data matrix according to function specified.
}
\description{
Calculates the specified normalized data matrix
}
