% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{getNormalizedCopy}
\alias{getNormalizedCopy}
\title{Calculates the specified normalized data matrix}
\usage{
getNormalizedCopy(data, func)
}
\arguments{
\item{data}{numeric matrix}

\item{func}{normalization method to apply}
}
\value{
Normalized data matrix according to function specified.
}
\description{
Calculates the specified normalized data matrix
}
