% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{getNormalizedCopySparse}
\alias{getNormalizedCopySparse}
\title{Calculates the specified normalized data matrix}
\usage{
getNormalizedCopySparse(data, func)
}
\arguments{
\item{data}{numeric matrix}

\item{func}{normalization method to apply}
}
\value{
NormData normalized data object that maintains sparse structure
}
\description{
Calculates the specified normalized data matrix
}
