% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getParam}
\alias{getParam}
\title{Gets parameters with defaults}
\usage{
getParam(object, param, subparam = NULL)
}
\arguments{
\item{object}{VISION object}

\item{param}{name of parameter that is desired}
}
\value{
bool whether or not there is unnormalize data
}
\description{
Used to get analysis parameters and supply defaults if needed
}
