% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{getProjections,Vision-method}
\alias{getProjections,Vision-method}
\alias{getProjections}
\title{Get 2D views of the expression data}
\usage{
\S4method{getProjections}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
List of matrix (Cells x 2)
}
\description{
This method provides access to the 2d projections that are used
to display results in the output report
}
\examples{
\dontrun{

# After running 'analyze'
# Retrieve tSNE30 (tSNE with perplexity 30) and plot it

tsne <- getProjections(vis)[["tSNE30"]]

plot(tsne[, 1], tsne[, 2])

# To see the names of available projections, just run:

names(getProjections(vis))

}
}
