% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\name{getSelections,Vision-method}
\alias{getSelections,Vision-method}
\alias{getSelections}
\title{Get saved selections}
\usage{
\S4method{getSelections}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
Named list of selections.  Each selection is a character vector of cell/pool IDs
}
\description{
Access saved groups of cell IDs defined while using the interactive output report
}
\details{
This method allows you to retrieve saved selections later in R for downstream analyses

Note:  In order for selections to correctly save when launching the report, the report
       must be run by storing the results back into the object.

E.g.
\preformatted{vis <- viewResults(vis)}
and not
\preformatted{viewResults(vis)}
}
\examples{
\dontrun{

vis <- viewResults(vis)  # Selections saved while viewing results

# Retrieve cell IDs for a selection group named 'interesting cells'
interestingCells <- getSelections(vis)[['interesting cells']]

}
}
