% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\name{getSignatureAutocorrelation,Vision-method}
\alias{getSignatureAutocorrelation,Vision-method}
\alias{getSignatureAutocorrelation}
\title{Get Signature Autocorrelation Scores}
\usage{
\S4method{getSignatureAutocorrelation}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
data.frame with columns 'C', 'pValue', and 'FDR'
}
\description{
Access the local autocorrelation scores computed for signatures
}
\details{
Local autocorrelation scores are calculated from the input latent
space (default's to PCA) or the input trajectory model (if provided)
}
