% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{getSignatureConsistency,Vision-method}
\alias{getSignatureConsistency,Vision-method}
\alias{getSignatureConsistency}
\title{Get Signature Consistency Scores}
\usage{
\S4method{getSignatureConsistency}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
data.frame with columns 'Consistency', 'pValue', and 'FDR'
}
\description{
Access the local consistency scores computed for signatures
}
\details{
Local consistency scores are calculated from the input latent
space (default's to PCA) or the input trajectory model (if provided)
}
