% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{getSignatureDifferential,Vision-method}
\alias{getSignatureDifferential,Vision-method}
\alias{getSignatureDifferential}
\title{Get Results of One-vs-All Differential Signature Tests}
\usage{
\S4method{getSignatureDifferential}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
nested list of list of data.frame (see details)
}
\description{
Returns the results of running one-vs-all differential signature
tests for each level of every factor meta-variable.
}
\details{
The 'stat' variable refers to the AUC

The output object has a nested structure:

List of meta-data variables -> List of variable levels -> Results Dataframe

The results dataframe has three columns: "stat", "pValue", "FDR"
}
