% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\name{getSignatureScores,Vision-method}
\alias{getSignatureScores,Vision-method}
\alias{getSignatureScores}
\title{Get Signature Scores}
\usage{
\S4method{getSignatureScores}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
Signature scores as a (Cells x Signature) matrix
}
\description{
Access to the signature scores computed by VISION
}
