% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{getVerts}
\alias{getVerts}
\title{Utility function for computing the closest vertex for each cell in a projection}
\usage{
getVerts(projection)
}
\arguments{
\item{projection}{TreeProjection object}
}
\value{
numeric: closest vertex for each cell in projection
}
\description{
Utility function for computing the closest vertex for each cell in a projection
}
