% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getWorkerCount}
\alias{getWorkerCount}
\title{Gets number of processes available for multicore work}
\usage{
getWorkerCount()
}
\value{
integer number of cores
}
\description{
Gets number of processes available for multicore work
}
