% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{group_modules_enrichment}
\alias{group_modules_enrichment}
\title{Make the clusters for the modules by enrichment.
For now we just assign each signature to each cluster, could filter to only include once,
so that each one appears in the modules x sigs table.}
\usage{
group_modules_enrichment(stats, pvals)
}
\arguments{
\item{stats}{overlap stats from calc_set_enrichment}

\item{pvals}{overlap p values from calc_set_enrichment}
}
\value{
assignments of each signature to each module
}
\description{
Make the clusters for the modules by enrichment.
For now we just assign each signature to each cluster, could filter to only include once,
so that each one appears in the modules x sigs table.
}
