% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{hasUnnormalizedData}
\alias{hasUnnormalizedData}
\title{Tests for Unnormalized Data}
\usage{
hasUnnormalizedData(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
bool whether or not there is unnormalize data
}
\description{
Determines if the VISION object is storing unnormalized data
}
