% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{Hotspot,Vision-method}
\alias{Hotspot,Vision-method}
\title{Run hotspot on this vision object.}
\usage{
\S4method{Hotspot}{Vision}(object, counts)
}
\arguments{
\item{object}{VISION object}

\item{counts}{}
}
\value{
VISION object
}
\description{
Run hotspot on this vision object.
}
