% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housekeeping}
\alias{housekeeping}
\title{Lists of housekeeping genes}
\format{A named list of character vectors
\describe{
  \item{default}{default list of housekeeping genes}
}}
\usage{
housekeeping
}
\description{
A dataset containing one or more lists of genes to be used as
 housekeeping genes when fitting the FNR curves.
}
\keyword{datasets}
