% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{hsCalculateModuleScores}
\alias{hsCalculateModuleScores}
\title{Create Hotspot Modules and calculate module scores given a HS object
with local correlations already calculated}
\usage{
hsCalculateModuleScores(
  hs,
  min_gene_threshold = 20,
  clustering_fdr = 0.5,
  plot = F
)
}
\arguments{
\item{hs}{the Hotspot object, must have ran compute_local_correlations already}

\item{min_gene_threshold}{min genes per module}

\item{clustering_fdr}{p value for clustering genes}
}
\value{
the modified hs object
}
\description{
Create Hotspot Modules and calculate module scores given a HS object
with local correlations already calculated
}
