% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{hsComputeAutoCorrelations}
\alias{hsComputeAutoCorrelations}
\title{Compute Hotspot auto correlations}
\usage{
hsComputeAutoCorrelations(
  hs,
  number_top_genes = 1000,
  autocorrelation_fdr = 0.05
)
}
\arguments{
\item{hs}{the Hotspot object}

\item{number_top_genes}{Hotspot argument for number of genes to consider}

\item{autocorrelation_fdr}{threshold for significance for genes autocorr}
}
\value{
list of HS genes
}
\description{
Compute Hotspot auto correlations
}
