% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{hsComputeLocalCorrelations}
\alias{hsComputeLocalCorrelations}
\title{Interface function to compute local correlations for Hotspot
Warning: modifies the hs argument}
\usage{
hsComputeLocalCorrelations(hs, hs_genes)
}
\arguments{
\item{hs}{the Hotspot object}

\item{hs_genes}{Hotspot genes}
}
\value{
the populated hs object
}
\description{
Interface function to compute local correlations for Hotspot
Warning: modifies the hs argument
}
