% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{hsCreateKnnGraph}
\alias{hsCreateKnnGraph}
\title{Init KNN graph in Hotspot object}
\usage{
hsCreateKnnGraph(hs, object, n_neighbors = NULL)
}
\value{
the Hotspot object with KNN initialized
}
\description{
Init KNN graph in Hotspot object
}
