% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{hsInit}
\alias{hsInit}
\title{Init Hotspot object from Vision Object}
\usage{
hsInit(object, model = "normal", tree = F, num_umi = NULL, logdata = FALSE)
}
\arguments{
\item{object}{the Vision Object}

\item{model}{the model for Hotspot (ie "normal", "danb"...)}

\item{tree}{boolean, whether to use the tree as ls}

\item{num_umi}{df of barcodes x num_umi}

\item{logdata}{boolean, log the expression data, avoid for danb}
}
\value{
the Hotspot object
}
\description{
Init Hotspot object from Vision Object
}
