% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ilog1p}
\alias{ilog1p}
\title{inverse log-scale transform a dense OR sparse matrix}
\usage{
ilog1p(spmat)
}
\arguments{
\item{spmat}{sparse Matrix}
}
\value{
logmat sparse Matrix
}
\description{
This avoids the creation of a dense intermediate matrix when
operating on sparse matrices
}
\details{
Performs result <- exp(spmat) - 1
}
