% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\docType{methods}
\name{initialize,ExpressionData-method}
\alias{initialize,ExpressionData-method}
\title{Initializes an ExpressionData object}
\usage{
\S4method{initialize}{ExpressionData}(.Object, data, ...)
}
\arguments{
\item{data}{expression data matrix}
}
\value{
ExpressionData object
}
\description{
Initializes an ExpressionData object
}
\examples{
expr <- readExprtoMatrix("data/expression_matrix.txt")
edata <- ExpressionMatrix(expr)
}
