% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\docType{methods}
\name{initialize,ServerExpression-method}
\alias{initialize,ServerExpression-method}
\title{Wrapper class for ExpressionData object for JSON.
Should not be called directly, instead use the `new` syntax}
\usage{
\S4method{initialize}{ServerExpression}(.Object, data, sample_labels,
  gene_labels)
}
\arguments{
\item{.Object}{an object}

\item{data}{Expression data}

\item{sample_labels}{Labels of samples in expression data}

\item{gene_labels}{Lables of genes in expression data}
}
\value{
ServerExpression object
}
\description{
Wrapper class for ExpressionData object for JSON.
Should not be called directly, instead use the `new` syntax
}
