% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\docType{methods}
\name{initialize,ServerSigProjMatrix-method}
\alias{initialize,ServerSigProjMatrix-method}
\title{Wrapper class for Signature Projection Matrix
Should not be called directly, instead use the `new` syntax}
\usage{
\S4method{initialize}{ServerSigProjMatrix}(.Object, data, proj_labels,
  sig_labels)
}
\arguments{
\item{.Object}{an object}

\item{data}{Signatrue Projection matrix as obtained from sigsVsProjections}

\item{proj_labels}{Projection names}

\item{sig_labels}{Names of signatures}
}
\value{
ServerSigProjMatrix object
}
\description{
Wrapper class for Signature Projection Matrix
Should not be called directly, instead use the `new` syntax
}
