% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TreeProjectionData.R
\docType{methods}
\name{initialize,TreeProjectionData-method}
\alias{initialize,TreeProjectionData-method}
\title{Wrapper for storing all relevant information for a given projection.}
\usage{
\S4method{initialize}{TreeProjectionData}(.Object, projections = NULL, keys,
  sigProjMatrix, pMatrix, sigClusters, treeScore)
}
\arguments{
\item{.Object}{an object}

\item{projections}{List of Projection objects to be stored}

\item{keys}{Sample names of expression data}

\item{sigProjMatrix}{Matrix storing the median consistency score per projection, signature pair}

\item{pMatrix}{Matrix storing the p values for each projection, signature pair}

\item{sigClusters}{a list of signature clusters}

\item{treeScore}{a significance score for the fitted tree}
}
\value{
ProjectionData object
}
\description{
Stores a list of Projection objects, filter name, and a logical value indicating whether or not
PCA was performed. Also stores clusters, a signature to projection matrix, and relevant gene names
and signature / projection keys.
Initializes a ProjectionData object for neatly storing all relevant data to a given projection section
Should not be called directly, instead use the `new` syntax
}
