% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{innerEvalSignatureBatch}
\alias{innerEvalSignatureBatch}
\title{Used in inner loop of batchSigEval}
\usage{
innerEvalSignatureBatch(exprData, sigs, weights = matrix(NA, 1, 1))
}
\arguments{
\item{exprData}{numeric Matrix Genes x Cells}

\item{sigs}{List of Signature to be evalauting}

\item{weights}{numeric Matrix Genes x Cells}
}
\value{
matrix containing signature values (sigs x cells)
}
\description{
Used in inner loop of batchSigEval
}
