% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{innerEvalSignatureBatchNorm}
\alias{innerEvalSignatureBatchNorm}
\title{Used in inner loop of batchSigEvalNorm}
\usage{
innerEvalSignatureBatchNorm(normData, sigs)
}
\arguments{
\item{normData}{NormData row/column normalization factors}

\item{sigs}{List of Signature to be evalauting}
}
\value{
matrix containing signature values (sigs x cells)
}
\description{
Computes signature scores without inflating the genes/cells matrix
}
