% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{is_tip}
\alias{is_tip}
\title{Check if a child is a tip}
\usage{
is_tip(tree, node)
}
\arguments{
\item{tree}{an object of class phylo}

\item{node}{the node to check}
}
\value{
true or false
}
\description{
Check if a child is a tip
}
