% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{knn_tree}
\alias{knn_tree}
\title{Helper KNN Function for Trees}
\usage{
knn_tree(leaves, k, distances)
}
\arguments{
\item{leaves}{leaves to find the nearest neighbors of}

\item{k}{number of neighbors to query}

\item{distances}{matrix of Samples x Samples matrix of cophenetic tree distances}
}
\value{
list with two items:
    index: Samples x K matrix of neighbor indices
    dist: Samples x K matrix of neighbor distances
}
\description{
Computes nearest-neighbor indices and distances
in serial.  Query is over all points and results
do not include the self-distance.
}
