% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{launchServer}
\alias{launchServer}
\title{Lanch the server}
\usage{
launchServer(object, port = NULL, host = NULL, browser = TRUE)
}
\arguments{
\item{object}{Vision object}

\item{port}{The port on which to serve the output viewer.  If omitted, a
random port between 8000 and 9999 is chosen.}

\item{host}{The host used to serve the output viewer. If omitted, "127.0.0.1"
is used.}

\item{browser}{Whether or not to launch the web browser}
}
\value{
object
}
\description{
Lanch the server
}
