% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{lca_based_depth}
\alias{lca_based_depth}
\title{Depth of tip1 parent immediately after LCA(tip1, tip2)}
\usage{
lca_based_depth(tree, tip1, tip2)
}
\arguments{
\item{tree}{an object of class phylo}

\item{tip1}{the first leaf}

\item{tip2}{the second leaf}
}
\value{
the trivial distance between tip1, tip2
}
\description{
Depth of tip1 parent immediately after LCA(tip1, tip2)
}
