% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{lcaBasedHotspotNeighbors}
\alias{lcaBasedHotspotNeighbors}
\title{Add custom tree based neighbor and weights to a Hotspot object}
\usage{
lcaBasedHotspotNeighbors(tree, hotspot, minSize = 20)
}
\description{
@param tree object of class phylo
 @param the Hotspot object to add the nw to
 @param minSize the minimum number of neighbors of the node
 @return the Hotspot object
 
 @export
}
