% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{lcaBasedTreeKNN}
\alias{lcaBasedTreeKNN}
\title{Generate neighbors and weights for a tree object, based on LCA.}
\usage{
lcaBasedTreeKNN(tree, minSize = 20)
}
\description{
@param tree object of class phylo
 @param minSize the minimum number of neighbors per node
 @return the hotspot object
}
