% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{loadHotspotObject}
\alias{loadHotspotObject}
\title{Load in an existing Hotspot object from bytes or a file}
\usage{
loadHotspotObject(file = NULL, bytes = NULL)
}
\arguments{
\item{file}{optional path to an existing file containing pickled bytes (format 0)}

\item{bytes}{optional R character vector of bytes created by calcHotspotModules}
}
\value{
an externalptr to a Hotspot Object loaded in the R reticulate session
}
\description{
Load in an existing Hotspot object from bytes or a file
}
