% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{louvainCluster}
\alias{louvainCluster}
\title{Applies the Louvain algorithm to generate micro-clustered data}
\usage{
louvainCluster(kn, data)
}
\arguments{
\item{kn}{List of nearest neighbor indices and euclidean distances to these nearest neighbors}

\item{data}{Data matrix}
}
\value{
List of clusters, each entry being a vector of indices representing samples in the cluster.
}
\description{
Applies the Louvain algorithm to generate micro-clustered data
}
