% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{matLog2}
\alias{matLog2}
\title{log2-scale transform a dense OR sparse matrix}
\usage{
matLog2(spmat, scale = FALSE, scaleFactor = 1e+06)
}
\arguments{
\item{spmat}{sparse Matrix}

\item{scale}{boolean - whether or not to scale the columns to sum to `scale_factor`}

\item{scaleFactor}{if scale = TRUE, columns are scaled to sum to this number}
}
\value{
logmat sparse Matrix
}
\description{
This avoids the creation of a dense intermediate matrix when
operating on sparse matrices
}
\details{
Either performs result <- log2(spmat+1) or if scale = TRUE
returns result <- log2(spmat/colSums(spmat)*scaleFactor + 1)
}
