% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{matrix_chisq}
\alias{matrix_chisq}
\title{Perform 1vAll factor analysis given a factor matrix and group definition}
\usage{
matrix_chisq(factorDF, cluster_ii)
}
\arguments{
\item{factorDF}{dataframe of factors}

\item{cluster_ii}{numeric vector - indices denoting the group to be compared}
}
\value{
pval - numeric vector, pvalue for each column

stat - numeric vector, test statistic (Cramer's V) for each column
}
\description{
Given indices in cluster_ii, compute the chisq test
on every column for values in cluster_ii vs the rest.
}
