% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{matrix_wilcox}
\alias{matrix_wilcox}
\title{Vectorized wilcox rank-sums test}
\usage{
matrix_wilcox(ranks, cluster_ii, check_na = FALSE, check_ties = FALSE)
}
\arguments{
\item{ranks}{matrix of ranks, each column representing a separate variable}

\item{cluster_ii}{numeric vector - indices denoting the group to be compared}

\item{check_na}{- whether or not to check for NA values (slows computation)}

\item{check_ties}{- whether or not to check for ties in ranks (slows computation)}
}
\value{
pval - numeric vector, pvalue for each column

stat - numeric vector, test statistic (AUC) for each column
}
\description{
Given indices in cluster_ii, compute the ranksums test
on every column for values in cluster_ii vs the rest.
}
