% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{matrix_wilcox_cpp}
\alias{matrix_wilcox_cpp}
\title{C++ wilcox rank-sums test}
\usage{
matrix_wilcox_cpp(data, cluster_num, cluster_denom,
  jobs = getOption("mc.cores", 1L))
}
\arguments{
\item{data}{matrix of values, each row representing a separate variable}

\item{cluster_num}{numeric vector - indices denoting the group to be compared}

\item{cluster_denom}{numeric vector - indices denoting the other group to be compared}

\item{jobs}{integer specifying number of parallel jobs.  Can be used to override
the default mc.cores option for when running this within an already-parallel loop}
}
\value{
pval - numeric vector, pvalue for each row

stat - numeric vector, test statistic (AUC) for each row
}
\description{
Given indices in cluster_num, and cluster_denom, compute the ranksums test
on every row for values in cluster_num vs cluster_denom.
}
\details{
Ties are handled correctly but values are assumed to contain no NAs
}
