% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{merge_intervals}
\alias{merge_intervals}
\title{merge overlapping intervals}
\usage{
merge_intervals(intervals)
}
\arguments{
\item{intervals}{a list of intervals}
}
\value{
a list of intervals where overlaps are merged into a single interval.
}
\description{
merge overlapping intervals
}
