% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{minSizeCladeNeighbors}
\alias{minSizeCladeNeighbors}
\title{Get's the nearest >= min size neighbors of a node based on clade structure}
\usage{
minSizeCladeNeighbors(tree, tip, minSize = 20)
}
\arguments{
\item{tree}{an object of class phylo}

\item{tip}{the tip to find the neighbors of}

\item{minSize}{the minimum number of neighbors of the node (excludes self)}
}
\value{
the neighbors
}
\description{
Get's the nearest >= min size neighbors of a node based on clade structure
}
