% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{naiveEvalSignature}
\alias{naiveEvalSignature}
\title{Different ways to evalutate the signature score}
\usage{
naiveEvalSignature(exprData, sig, weights)
}
\arguments{
\item{exprData}{numeric Matrix Genes x Cells}

\item{sig}{Signature to be evalauting}

\item{weights}{Weight matrix computed through FNR curve}
}
\value{
SignatureScore object
}
\description{
EAch method should have the same signature so they can be swapped
}
\details{
Right now, each takes in a wrapped data object and signature object

Specified by FastProject argument (sig_score_method), default = naiveEvalSignature
Naive eval Siganture, just sums the columns * sign, equivalent to all weights = 1
}
