% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{newAnalysis}
\alias{newAnalysis}
\title{Run the analysis again wth user-defined subsets or confguration}
\usage{
newAnalysis(nfp)
}
\arguments{
\item{nfp}{the new FastProject object to analyze}
}
\description{
Run the analysis again wth user-defined subsets or confguration
}
