% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{noNormalization}
\alias{noNormalization}
\title{Does nothing, just returns the original data}
\usage{
noNormalization(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
The data matrix
}
\description{
Does nothing, just returns the original data
}
