% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize the data}
\usage{
normalizeData(object, sig_norm_method = object@sig_norm_method)
}
\arguments{
\item{object}{the FastPrpject object}

\item{sig_norm_method}{(optional) Method to apply to normalize the expression
matrix before calculating signature scores}
}
\value{
the FastProject object with normalized data matrices
}
\description{
Normalize the data
}
