% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{NormData}
\alias{NormData}
\title{Initialize a new NormData object}
\usage{
NormData(
  data,
  rowOffsets = NULL,
  rowScaleFactors = NULL,
  colOffsets = NULL,
  colScaleFactors = NULL
)
}
\arguments{
\item{data}{expression data matrix}

\item{rowOffsets}{offsets to be subtracted from each row}

\item{rowScaleFactors}{factors to scale each row}

\item{colOffsets}{offsets to be subtracted from each column}

\item{colScaleFactors}{factors to scale each column}
}
\value{
NormData object
}
\description{
Initialize a new NormData object
}
