% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{pearsonCorrToJSON}
\alias{pearsonCorrToJSON}
\title{convert perason correlation coeffcients between PCs and sgnatures into a JSON object}
\usage{
pearsonCorrToJSON(pc, sigs)
}
\arguments{
\item{pc}{the pearson correlation coefficients matrix}

\item{sigs}{the signatures of interest}
}
\value{
Subsetted pearson correlations converted to JSON
}
\description{
convert perason correlation coeffcients between PCs and sgnatures into a JSON object
}
