% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\name{phyloAnalyze,PhyloVision-method}
\alias{phyloAnalyze,PhyloVision-method}
\alias{phyloAnalyze}
\title{Analyze a PhyloVision object}
\usage{
\S4method{phyloAnalyze}{PhyloVision}(object, hotspot = FALSE)
}
\arguments{
\item{...}{arguments passed to the base Vision constructor}
}
\description{
Analyze a PhyloVision object
}
