% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Vision.R
\name{PhyloVision}
\alias{PhyloVision}
\alias{PhyloVision,phylo-method}
\title{Initializes a new PhyloVision Object}
\usage{
PhyloVision(tree, ...)

\S4method{PhyloVision}{phylo}(tree, ...)
}
\arguments{
\item{tree}{parsed ape tree}

\item{...}{arguments passed to the base Vision constructor}
}
\description{
Initializes a new PhyloVision Object
}
