% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{poolCells}
\alias{poolCells}
\title{create micro-clusters that reduce noise and complexity while maintaining
the overall signal in the data}
\usage{
poolCells(object)
}
\arguments{
\item{object}{the FastProject object to cluster the cells of}
}
\value{
the FastProject with pooled cells
}
\description{
create micro-clusters that reduce noise and complexity while maintaining
the overall signal in the data
}
