% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{poolMatrixCols}
\alias{poolMatrixCols}
\title{Pools columns of a numeric matrix}
\usage{
poolMatrixCols(data, pools)
}
\arguments{
\item{data}{data.frame or matrix to pool}

\item{pools}{named list of character vector describing cell ids
in each pool}
}
\value{
Matrix of pooled data
}
\description{
Uses the provided pools to merge columns of the supplied data matrix
}
\details{
This would typically be used on a gene expression matrix (genes X cells) to
pool cells.

The \code{pools} argument is obtained by running \code{applyMicroClustering}
}
