% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{poolMatrixRows}
\alias{poolMatrixRows}
\title{Pools rows of a numeric matrix}
\usage{
poolMatrixRows(data, pools)
}
\arguments{
\item{data}{data.frame or matrix to pool}

\item{pools}{named list of character vector describing cell ids
in each pool}
}
\value{
Matrix of pooled data
}
\description{
Uses the provided pools to merge rows of the supplied data matrix
}
\details{
Same as poolMatrixCols only operates on rows.

The \code{pools} argument is obtained by running \code{applyMicroClustering}
}
