% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{processSignatures}
\alias{processSignatures}
\title{Processes signatures on input}
\usage{
processSignatures(sigData, exprData, minSignatureGenes, sig_gene_threshold)
}
\arguments{
\item{sigData}{list of signature objects}

\item{exprData}{gene expression matrix}

\item{minSignatureGenes}{minimum number of genes a signature must match
in the expression matrix in order to be retained}

\item{sig_gene_threshold}{Proportion of cells that a gene must be detected in (nonzero)
to be used in signature score calculations.}
}
\value{
processedSigData list of signature objects
}
\description{
Removes genes that don't match a gene in the expression matrix
Drops signatures with less than `minSignatureGenes` matching genes
}
