% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ProjectionData.R
\name{ProjectionData}
\alias{ProjectionData}
\title{Wrapper for storing all relevant information for a given consistency test}
\usage{
ProjectionData(Consistency, pValue, FDR, sigClusters)
}
\arguments{
\item{Consistency}{Matrix storing the consistency score per signature pair}

\item{pValue}{Matrix storing the p values for each projection, signature pair}

\item{FDR}{Matrix storing the FDR for each p-value (BH method)}

\item{sigClusters}{the signature clusters}
}
\value{
ProjectionData object
}
\description{
Stores a list of consistency test results
Also stores clusters, a signature to projection matrix
Initializes a ProjectionData object for storing the consistency test results
}
