% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBasedMethods.R
\name{projectOnTree}
\alias{projectOnTree}
\title{Project the given dataoints onto the tree defined by the vertices (V.pos) and binary adjacency matrix (princAdj)}
\usage{
projectOnTree(data.pnts, V.pos, princAdj)
}
\arguments{
\item{data.pnts}{(D x N numeric) the spatial coordinates of data points to project}

\item{V.pos}{(D x K numeric) the spatial coordinates of the tree vertices}

\item{princAdj}{(K x K logical) a symmetric binary adjacency matrix (K x K)}
}
\value{
(list) projection information:
  \itemize{
    \item{"spatial"}{The D-dimensional position of the projected data points}
    \item{"edge"}{a Nx2 matrix, were line i has the indices identifying the edge that datapoint i was projected on,
    represented as (node a, node b). For consistency and convenience, it is maintained that a < b}
    \item{"edgePos"}{an N-length numeric with values in [0,1], the relative position on the edge of the datapoint.
    0 is node a, 1 is node b, .5 is the exact middle of the edge, etc.}
 }
}
\description{
Project the given dataoints onto the tree defined by the vertices (V.pos) and binary adjacency matrix (princAdj)
}
\details{
data points are projected on their nearest edge, defined to be the edge that is connected to the nearest node
and has minimal length of orthogonal projection. Data points are projected with truncated orthogonal projection:
point that fall beyond the edge are projected to the closer node.
}
\examples{
X <- matrix(rnorm(200), nrow = 2)
tree <- applySimplePPT(X)
proj <- projectOnTree(X, tree[[1]], tree[[2]])
}
