% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{PseudotimeProjection}
\alias{PseudotimeProjection}
\title{Computes pseudotime projection for a given TreeProjection and root.}
\usage{
PseudotimeProjection(projection, root)
}
\arguments{
\item{projection}{TreeProjection object}

\item{root}{cell to use for the beginning of the pseudotime projection}
}
\value{
list:
\itemize{
    \item path i: a list containing the cells and pseudotime coordinates
    for each path originating from root.
}
}
\description{
Computes pseudotime projection for a given TreeProjection and root.
}
