% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{ptSling.to.ptLin}
\alias{ptSling.to.ptLin}
\title{Utility function converting slingshot trajecories to LineagePulse input
arguments.}
\usage{
ptSling.to.ptLin(ptSling)
}
\arguments{
\item{ptSLing}{matrix numeric cells X curves}
}
\value{
data.frame: index cells columns curves
}
\description{
Utility function converting slingshot trajecories to LineagePulse input
arguments.
}
