% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\docType{data}
\name{RANDOM_SEED}
\alias{RANDOM_SEED}
\title{Contains information that is useful to share
across all modules.}
\format{An object of class \code{numeric} of length 1.}
\usage{
RANDOM_SEED
}
\description{
Contains information that is useful to share
across all modules.
}
\keyword{datasets}
