% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{read_10x}
\alias{read_10x}
\title{Read 10x Output}
\usage{
read_10x(expression, genes, barcodes, ensToSymbol = TRUE)
}
\arguments{
\item{expression}{path to matrix.mtx}

\item{genes}{path to genes.tsv}

\item{barcodes}{path to barcodes.tsv}

\item{ensToSymbol}{bool denoting whether or not to perform label conversion}
}
\value{
sparse count matrix with appropriate row/column names
}
\description{
Loads 10x output counts and converts expression to gene symbols
}
\details{
This version takes in three files as inputs:
\enumerate{
  \item matrix.mtx
  \item genes.tsv
  \item barcodes.tsv
}

These files are found in the output of "cellranger count" in a folder
that looks like:

\code{outs/filtered_gene_bc_matrices/mm10}

though with the name of whichever genome you are using instead of 'mm10'
}
