% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{read_10x_h5}
\alias{read_10x_h5}
\title{Read 10x HDF5 Output}
\usage{
read_10x_h5(h5_file, ensToSymbol = TRUE)
}
\arguments{
\item{h5_file}{path to h5 file}

\item{ensToSymbol}{bool denoting whether or not to perform label conversion}
}
\value{
Return value depends on whether this is a Cellranger v3 or Cellranger v2 output
    If it is a v2 output, return is a sparse count matrix with gene expression values
    If it is a v3 output, return value is a list with two entries:
        Expression: sparse count matrix with gene expression counts (genes x cells)
        Antibody: sparse count matrix with antibody capture counts (cells x antibodies)
}
\description{
Loads 10x output counts and converts expression to gene symbols
}
\details{
This version uses the h5 file produced by "cellranger count"

This file is typically in a folder that looks like:

\code{outs/filtered_gene_bc_matrices_h5.h5}
}
