% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readExprAsMatrix}
\alias{readExprAsMatrix}
\title{Reads in an expression matrix from a file path to a matrix}
\usage{
readExprAsMatrix(filename, sep = "\\t")
}
\arguments{
\item{filename}{path to the file to be read in}

\item{sep}{seperator to use while reading in the file. Default is "\\t"}
}
\value{
Matrix read in from filename.
}
\description{
Reads in an expression matrix from a file path to a matrix
}
