% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\docType{methods}
\name{readExprData,ExpressionData-method}
\alias{readExprData,ExpressionData-method}
\title{Prints out expression data}
\usage{
\S4method{readExprData}{ExpressionData}(object)
}
\arguments{
\item{object}{ExpressionData object}
}
\value{
Nothing.
}
\description{
Prints out expression data
}
\examples{
expr <- readExprtoMatrix("data/expression_matrix.txt")
edata <- ExpressionMatrix(expr)
readExprData(eData) 
}
