% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readHKGToMatrix}
\alias{readHKGToMatrix}
\title{Reads in a .txt file containing Housekeeping Gene names}
\usage{
readHKGToMatrix(filename, sep = "\\t")
}
\arguments{
\item{filename}{path to the file to be read in}

\item{sep}{seperator to use while reading in the file. Default is tab.}
}
\value{
Table containing the houseekeeping gene names
}
\description{
Reads in a .txt file containing Housekeeping Gene names
}
