% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{readjust_clusters}
\alias{readjust_clusters}
\title{Repartitions existing clusters to achieve desired granularity.}
\usage{
readjust_clusters(clusters, data, cellsPerPartition = 100)
}
\arguments{
\item{clusters}{List of clusters, each entry being a vector of cells in a
cluster.}

\item{data}{NUM_SAMPLES x NUM_PROTEINS data matrix that was used to generate
clusters}

\item{cellsPerPartition}{the number of cells for a single partition the
algorithm should aim for}
}
\value{
Repartitioned clusters, such that a desireable number of
microclusters is acheived.
}
\description{
By default, minimum number of clusters to be generated is the squareroot of
the number of cells.
}
