% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readSignaturesInput}
\alias{readSignaturesInput}
\title{Reads in a list of signature input files.}
\usage{
readSignaturesInput(filenames)
}
\arguments{
\item{filenames}{a list of paths to signature input files}
}
\value{
List of Signature Objects
}
\description{
Reads in a list of signature input files.
}
\examples{
sigs <- readSignaturesInput(c("sigs1.txt", "sigs2.gmt"))
}
