% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{registerMethods}
\alias{registerMethods}
\title{Registers the projection methods to be used}
\usage{
registerMethods(projection_methods, numCells)
}
\arguments{
\item{projection_methods}{character vector of projection methods to use}

\item{numCells}{number of cells in this analysis}
}
\value{
List of projection methods to be applied.
}
\description{
Registers the projection methods to be used
}
