% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{rowAndColNormalization}
\alias{rowAndColNormalization}
\title{Performs z-normalization on all columns and rows}
\usage{
rowAndColNormalization(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
Data matrix with same dimensions, with each row and column z-normalized.
}
\description{
Performs z-normalization on all columns and rows
}
