% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{rowNormalization}
\alias{rowNormalization}
\title{Performs z-normalization on all rows}
\usage{
rowNormalization(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
Data matrix with same dimensions, with each row z-normalized.
}
\description{
Performs z-normalization on all rows
}
