% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{rowVarsSp}
\alias{rowVarsSp}
\title{Compute row-wise variance on matrix without densifying}
\usage{
rowVarsSp(x)
}
\arguments{
\item{x}{expression matrix}
}
\value{
numeric vector row-wise variance
}
\description{
Compute row-wise variance on matrix without densifying
}
