% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{runHotspot}
\alias{runHotspot}
\title{Perform Hotspot analysis on Vision Object}
\usage{
runHotspot(
  object,
  model = "normal",
  tree = FALSE,
  number_top_genes = 1000,
  num_umi = NULL,
  min_gene_threshold = 20,
  n_neighbors = NULL,
  autocorrelation_fdr = 0.05,
  clustering_fdr = 0.5,
  logdata = FALSE
)
}
\arguments{
\item{object}{Vision Object}

\item{model}{model argument for Hotspot, one of \itemize{
\item normal
\item danb
\item bernoulli
\item none 
}}

\item{tree}{whether to use tree as latent space. If TRUE, object should have
a tree slot.}

\item{number_top_genes}{Hotspot argument for number of genes to consider}

\item{num_umi}{optional dataframe containing umi counts in first column for
barcodes}

\item{min_gene_threshold}{minimum number of genes in Hotspot module}

\item{n_neighbors}{number of neighbors to consider in latent space}

\item{autocorrelation_fdr}{threshold for significance for genes autocorr}

\item{clustering_fdr}{threshold for significance for clustering modules}

\item{logdata}{boolean, log the expression data, avoid for danb
Populates the modData, HotspotModuleScores, ModuleSignatureEnrichment
and HotspotObject slots of object, as well as recalculates signature scores
for new modules.}
}
\value{
the modified Vision object
}
\description{
Perform Hotspot analysis on Vision Object
}
