% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\docType{methods}
\name{saveAndViewResults,FastProject-method}
\alias{saveAndViewResults,FastProject-method}
\alias{saveAndViewResults}
\title{Save the FastProjectOutput object as an .RDS file and view the results on a
localhost}
\usage{
\S4method{saveAndViewResults}{FastProject}(fpout, ofile = NULL)
}
\arguments{
\item{fpout}{FastProject object}

\item{ofile}{the path to save the object in. If NULL, the object is saved
in the working directory [default:NULL]}
}
\value{
the path of the saved file
}
\description{
Save the results object as an RDS file for future use, and launch a local
server to explore the results with a browser.
}
\examples{
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

# create housekeeping genes
hkg <- paste0("gene",sample(1:500, 50))

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- FastProject(data = expMat,
                     housekeeping = hkg,
                     signatures = sigs)

## Analyze requires actual non-random data to run properly
\dontrun{
fp.out <- Analyze(fp)
saveAndViewResults(fp.out)
}
}
