% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{saveAndViewResults,Vision-method}
\alias{saveAndViewResults,Vision-method}
\alias{saveAndViewResults}
\title{Save the VISION object as an .RDS file and view the results on a
localhost}
\usage{
\S4method{saveAndViewResults}{Vision}(object, ofile = NULL, port = NULL,
  host = NULL, browser = TRUE, name = NULL)
}
\arguments{
\item{object}{VISION object}

\item{ofile}{the path to save the object in. If NULL, the object is saved
in the working directory [default:NULL]}

\item{port}{The port on which to serve the output viewer.  If omitted, a
random port between 8000 and 9999 is chosen.}

\item{host}{The host used to serve the output viewer. If omitted, "127.0.0.1"
is used.}

\item{browser}{Whether or not to launch the browser automatically (default=TRUE)}

\item{name}{Name for the sample - is shown at the top of the output report}
}
\value{
the path of the saved file
}
\description{
This is just a convience function wrapping two function calls
}
\details{
saveAndViewResults(vis, 'vision_results.rds')

is equivalent to:

saveRDS(vis, 'vision_results.rds')
    viewResults(vis)
}
\examples{
\dontrun{

vis <- Vision(data = expMat, signatures = sigs)

options(mc.cores=10) # Use 10 cores
vis <- analyze(vis)

saveAndViewResults(vis, 'vision_results.rds') # Saves and launches dynamic output report
}
}
