% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProjectOutput.R
\docType{methods}
\name{saveFPOutAndViewResults,FastProjectOutput-method}
\alias{saveFPOutAndViewResults,FastProjectOutput-method}
\title{Save the FastProjectOutput object as an .RDS file and view the results on a localhost}
\usage{
\S4method{saveFPOutAndViewResults}{FastProjectOutput}(fpout)
}
\arguments{
\item{fpout}{FastProjectOutput object}
}
\value{
None
}
\description{
Save the FastProjectOutput object as an .RDS file and view the results on a localhost
}
\examples{
fpout <- Analyze(fp)
saveFPOutAndViewResults(fpout)
}
