% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Module.R
\name{saveHSBytestToPickle}
\alias{saveHSBytestToPickle}
\title{Save bytes in the Hotspot object slot to a file}
\usage{
saveHSBytestToPickle(path, bytes)
}
\arguments{
\item{path}{the file path}

\item{bytes}{the raw bytes, like in the Hotspot slot of a VISION Object}
}
\description{
Save bytes in the Hotspot object slot to a file
}
