% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_version_current}
\alias{schema_version_current}
\title{Get current schema version}
\usage{
schema_version_current()
}
\value{
Character scalar "0.1.0"
}
\description{
Returns the current schema version supported by vision.
}
\examples{
schema_version_current()
}
