% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{ServerPCorr}
\alias{ServerPCorr}
\title{Wrapper class for server Perason correlation data
Should not be called directly, instead use the `new` syntax}
\usage{
ServerPCorr(data, proj_labels, sig_labels)
}
\arguments{
\item{data}{pearson r correlation coefficients}

\item{proj_labels}{the labels of the projections (columns) in the data}

\item{sig_labels}{the labels of the signatures (rows) in the data}
}
\value{
a ServerPCorr object
}
\description{
Wrapper class for server Perason correlation data
Should not be called directly, instead use the `new` syntax
}
