% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{ServerPMatrix}
\alias{ServerPMatrix}
\title{Wrapper class for the P value matrix calculated during sigsVsProjections}
\usage{
ServerPMatrix(data, proj_labels, sig_labels)
}
\arguments{
\item{data}{P values for each signature, projection pair in the form of a matrix}

\item{proj_labels}{Projection names}

\item{sig_labels}{Names of signatures}
}
\value{
ServerPMatrix object
}
\description{
Wrapper class for the P value matrix calculated during sigsVsProjections
}
