% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{ServerSigProjMatrix}
\alias{ServerSigProjMatrix}
\title{Wrapper class for Signature Projection Matrix
Should not be called directly, instead use the `new` syntax}
\usage{
ServerSigProjMatrix(data, proj_labels, sig_labels)
}
\arguments{
\item{data}{Signatrue Projection matrix as obtained from sigsVsProjections}

\item{proj_labels}{Projection names}

\item{sig_labels}{Names of signatures}
}
\value{
ServerSigProjMatrix object
}
\description{
Wrapper class for Signature Projection Matrix
Should not be called directly, instead use the `new` syntax
}
