% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\docType{methods}
\name{sigEqual,Signature-method}
\alias{sigEqual,Signature-method}
\title{Method for testing whether or not signatures are equal to one another (if all signs, genes and name are the same.)}
\usage{
\S4method{sigEqual}{Signature}(object, compareSig)
}
\arguments{
\item{object}{Signature object}

\item{compareSig}{Other signature to compare against}
}
\value{
TRUE if sigs are equal, else FALSE
}
\description{
Method for testing whether or not signatures are equal to one another (if all signs, genes and name are the same.)
}
