% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{Signature}
\alias{Signature}
\title{Initialize a new Signature object.
Should not be called directly, instead use the `new` syntax}
\usage{
Signature(sigDict, name, source, metaData = "")
}
\arguments{
\item{sigDict}{Named list of signs for each gene in the signature}

\item{name}{Name of the signature}

\item{source}{File from which this signature was read from}

\item{metaData}{Metadata pertinent to signature}
}
\value{
Signature object
}
\description{
Initialize a new Signature object.
Should not be called directly, instead use the `new` syntax
}
