% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{sigProjMatrixPToJSON}
\alias{sigProjMatrixPToJSON}
\title{Converts the -log10(pvalues) of the consistency scores into a JSON object}
\usage{
sigProjMatrixPToJSON(sigpmp, sigs)
}
\arguments{
\item{sigpmp}{SigProjMatrix p values}

\item{sigs}{Signatrues to subset from sigpmp}
}
\value{
Subsetted sigProjMatrix_P converted to JSON
}
\description{
Converts the -log10(pvalues) of the consistency scores into a JSON object
}
