% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{sigRanksToJSON}
\alias{sigRanksToJSON}
\title{Converts list of signature ranks to JSON}
\usage{
sigRanksToJSON(ss)
}
\arguments{
\item{ss}{List of rank values}
}
\value{
Signature ranks as JSON, with names of each entry that of list names
}
\description{
Converts list of signature ranks to JSON
}
