% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SignatureScores.R
\name{SigScoresFromDataframe}
\alias{SigScoresFromDataframe}
\title{Generate signature scores based on an input data.frame}
\usage{
SigScoresFromDataframe(df, sampleLabels)
}
\arguments{
\item{df}{a data.frame object where each column is a cell signature. Names of
columns are signatue names. Rownames must correspond to appropriate samples}

\item{sampleLabels}{a character vector representing sample names in
the expression matrix}
}
\value{
a list of SignatureScore objects, one for each column in the
data.frame
}
\description{
These are either factors or numeric values with known values for each sample
Can be conditions of interest, batch annotations, qc information, etc.
}
\note{
Factor signatures and numeric signatures are treated differently. Make
sure that the class of each input column is the correct one.
}
