% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{sigScoresToJSON}
\alias{sigScoresToJSON}
\title{Converts row of sigantures score matrix to JSON}
\usage{
sigScoresToJSON(ss)
}
\arguments{
\item{ss}{List of signature scores}
}
\value{
Signature scores list to JSON, with names of each entry that of the list names
}
\description{
Converts row of sigantures score matrix to JSON
}
