% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{sigsToSparseMatrix}
\alias{sigsToSparseMatrix}
\title{Utility method to load signatures into a sparse matrix}
\usage{
sigsToSparseMatrix(sigs, expression)
}
\arguments{
\item{sigs}{List of Signature to be evalauting}

\item{expression}{numeric Matrix Genes x Cells}
}
\value{
CsparseMatrix containing signature values
}
\description{
Utility method to load signatures into a sparse matrix
}
