% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{sigsVsProjection_pcf}
\alias{sigsVsProjection_pcf}
\title{Evaluates the significance of each meta data factor signature vs. a
single projections weights}
\usage{
sigsVsProjection_pcf(sigScoresData, weights)
}
\arguments{
\item{sigScoresData}{List of SignatureScores Object, mapping signature
names to their value at each coordinate}

\item{weights}{numeric matrix of dimension N_SAMPLES x N_SAMPLES}
}
\value{
list:
\itemize{
    \item consistency: consistency scores
    \item pvals: pvalues
}
}
\description{
Evaluates the significance of each meta data factor signature vs. a
single projections weights
}
