% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{sigsVsProjection_pcn}
\alias{sigsVsProjection_pcn}
\title{Evaluates the significance of each meta data numeric signature vs. a
single projections weights}
\usage{
sigsVsProjection_pcn(metaData, weights, cells = NULL)
}
\arguments{
\item{metaData}{data.frame of meta-data for cells}

\item{weights}{numeric matrix of dimension N_SAMPLES x N_SAMPLES}

\item{cells}{list of cell names.  Subsets anlysis to provided cells.
If omitted, use all cells.}
}
\value{
list:
\itemize{
    \item consistency: consistency scores
    \item pvals: pvalues
}
}
\description{
Evaluates the significance of each meta data numeric signature vs. a
single projections weights
}
