% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{sigsVsProjections}
\alias{sigsVsProjections}
\title{Evaluates the significance of each signature vs. each projection.}
\usage{
sigsVsProjections(projections, sigScoresData, metaData, randomSigData)
}
\arguments{
\item{projections}{Maps projections to their spatial coordinates for each
sample}

\item{sigScoresData}{numeric matrix of signature scores
size is cells x signatures}

\item{metaData}{data.frame of meta-data for cells}

\item{randomSigData}{A list with two items:

  randomSigs: a list of signature score matrices.  Each list item
    represents permutation signatures generated for a specific size/balance,
    and is a numeric matrix of size cells X signatures

  sigAssignments: named factor vector assigning signatures to random background
    groups}
}
\value{
list:
\itemize{
    \item sigProbMatrix: the matrix of signature-projection consistency scores
    \item pVals: pvalues for the scores
}
}
\description{
Evaluates the significance of each signature vs. each projection.
}
