% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateBranchDataSet.R
\name{simBranchingDataSet}
\alias{simBranchingDataSet}
\title{Simulated bracnhing Trajectory.}
\usage{
simBranchingDataSet(scaNCellsPerEdge, scaNGroups, scaNConst, scaNLin,
  scaMumax = 100, scaSDMuAmplitude = 1, scaNSigsPerGroup = 5,
  vecNormConstExternal = NULL, vecDispExternal = NULL,
  vecGeneWiseDropoutRates = NULL, matDropoutModelExternal = NULL)
}
\value{
list:
\itemize{
    \item path i: a list containing the cells and pseudotime coordinates
    for each path originating from root.
}
}
\description{
Based on LineagePulse (Fischer 2017) simulateContinuousDataSet()
}
\details{
To be completed
}
