% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubSample.R
\name{splitSamples}
\alias{splitSamples}
\title{This module handles the splitting and joining of the data if sub-sampling is enabled
randomally subsample the data}
\usage{
splitSamples(data, sampleSize)
}
\arguments{
\item{data}{the data to subsample}

\item{sampleSize}{the size of the subsample}
}
\value{
a list:
\itemize{
    \item the data not included in the subsample
    \item the data included in the subsample
}
}
\description{
This module handles the splitting and joining of the data if sub-sampling is enabled
randomally subsample the data
}
