% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{sqdist}
\alias{sqdist}
\title{Alternative computation of distance matrix, based on matrix multiplication.}
\usage{
sqdist(X, Y)
}
\arguments{
\item{X}{n x d matrix}

\item{Y}{m x d matrix}
}
\value{
n x m distance matrix
}
\description{
Alternative computation of distance matrix, based on matrix multiplication.
}
