% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{testPath}
\alias{testPath}
\title{Utility function for running LineagePulse with default arguments}
\usage{
testPath(path, counts)
}
\arguments{
\item{path}{data.frame containing cells names and continous pseudotime}

\item{counts}{matrix genes X cells}
}
\value{
numeric: closest vertex for given cell in projection
}
\description{
Utility function for running LineagePulse with default arguments
}
