% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{testProjectionLineages}
\alias{testProjectionLineages}
\title{Computes path statistics for a TreeProjection starting from a given
root or search through all vertices}
\usage{
testProjectionLineages(projection, root = NULL, k = 0)
}
\arguments{
\item{projection}{TreeProjection object}

\item{root}{cell to use for the beginning of the pseudotime projection
default: NULL}

\item{k}{numeric max number of vertices in path. k=0 => full paths
default: 0}
}
\value{
list:
\itemize{
    \item path i: a list containing the cells and pseudotime coordinates
    for each path originating from root.
}
}
\description{
Computes path statistics for a TreeProjection starting from a given
root or search through all vertices
}
