% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{testSigGenes}
\alias{testSigGenes}
\title{Filter Lineages}
\usage{
testSigGenes(lsObjLP)
}
\arguments{
\item{lsObjLP}{list LineagePUlse objects}
}
\value{
list: containg the signifigant genes corresponding to LineagePulse obj
\itemize{
    \item sigGenes i
}
}
\description{
Filter Lineages
}
