% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Trajectory.R
\name{Trajectory}
\alias{Trajectory}
\title{Initialize a new Trajectory object.}
\usage{
Trajectory(input)
}
\arguments{
\item{input}{trajectory to model cell progression.  Wrapped result
of a trajectory inference by the dynverse/dynwrap library}
}
\value{
Trajectory object
}
\description{
Initialize a new Trajectory object.
}
