% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TrajectoryProjection.R
\name{TrajectoryProjection}
\alias{TrajectoryProjection}
\title{Initialize a new TrajectoryProjection object.}
\usage{
TrajectoryProjection(name, vData, pData, adjMat)
}
\arguments{
\item{name}{label for this projection}

\item{vData}{#Milestones x 2 matrix with coordinates for milestones}

\item{pData}{#Cells x 2 matrix with coordinates for cells}

\item{adjMat}{#Milestones x #Milestones adjacency matrix (0/1)}
}
\value{
TrajectoryProjection object
}
\description{
This represents a 2d view of data on a projection
}
