% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Trajectory.R
\name{translateCellPositions}
\alias{translateCellPositions}
\title{Translate cell positions}
\usage{
translateCellPositions(progressions, vData, edges)
}
\arguments{
\item{progressions}{data.frame describing cell positions between milestones}

\item{vData}{Mx2 matrix mapping miletones into 2d}

\item{edges}{Edges x 2 matrix describing connectivity between edges}
}
\value{
pData Cells x 2 matrix with cell positions in 2d
}
\description{
Maps cell positions between edges
}
