% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{treeClusterMinCladeSize}
\alias{treeClusterMinCladeSize}
\title{Generate clade-clusters for a tree of minimum size (unless children of root)}
\usage{
treeClusterMinCladeSize(tree, minSize = 30)
}
\arguments{
\item{tree}{object of class phylo}

\item{minSize}{minimum clade size for a clade to be expanded}
}
\value{
List of clusters, each entry being a vector of tips representing
WARNING: This won't work well for tree's with broad multifurcations
}
\description{
Generate clade-clusters for a tree of minimum size (unless children of root)
}
