% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TreeProjection.R
\name{TreeProjection}
\alias{TreeProjection}
\title{Create a TreeProjection object
Should not be called directly, instead use the `new` syntax}
\usage{
TreeProjection(pData, name, vData, adjMat)
}
\arguments{
\item{pData}{the projected coordinates of the data}

\item{name}{the name of the projection}

\item{vData}{the projected coordinates of the tree nodes}

\item{adjMat}{the adjacency matrix of the tree}
}
\value{
a TreeProjection object
}
\description{
Create a TreeProjection object
Should not be called directly, instead use the `new` syntax
}
