% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TreeProjectionData.R
\name{TreeProjectionData}
\alias{TreeProjectionData}
\title{Wrapper for storing all relevant information for a given projection.}
\usage{
TreeProjectionData(latentTree, projections, sigProjMatrix, pMatrix, sigClusters,
  treeScore, emp_pMatrix)
}
\arguments{
\item{projections}{List of Projection objects to be stored}

\item{sigProjMatrix}{Matrix storing the median consistency score per projection, signature pair}

\item{pMatrix}{Matrix storing the p values for each projection, signature pair}

\item{sigClusters}{a list of signature clusters}

\item{treeScore}{a significance score for the fitted tree}

\item{emp_pMatrix}{Matrix storing the empirical p values for each projection, signature pair}
}
\value{
ProjectionData object
}
\description{
Stores a list of Projection objects.
Also stores clusters, a signature to projection matrix
Initializes a ProjectionData object for neatly storing all relevant data to a given projection section
Should not be called directly, instead use the `new` syntax
}
