% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{trivial_dist}
\alias{trivial_dist}
\title{Trivial distance function for arbitrary tree clustering}
\usage{
trivial_dist(tree, tip1, tip2)
}
\arguments{
\item{tree}{an object of class phylo}

\item{tip1}{the first leaf}

\item{tip2}{the second leaf}
}
\value{
the trivial distance between tip1, tip2
}
\description{
Number of mutations along path from tip1 to LCA(tip1, tip2)
Ensures if on same clade, join.
}
