% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ultrametric_tree}
\alias{ultrametric_tree}
\title{Generate an ultrametric tree}
\usage{
ultrametric_tree(tree)
}
\arguments{
\item{tree}{an object of class phylo}
}
\value{
a tree with edge distances such that it is ultrametric.
}
\description{
Generate an ultrametric tree
}
