% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\docType{methods}
\name{updateExprData,ExpressionData-method}
\alias{updateExprData,ExpressionData-method}
\title{Updates the expression data stored in this object}
\usage{
\S4method{updateExprData}{ExpressionData}(object, newData)
}
\arguments{
\item{object}{ExpressionData object}

\item{newData}{new expression data matrix}
}
\value{
ExpressionData object with updated data
}
\description{
Updates the expression data stored in this object
}
