% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Projection.R
\docType{methods}
\name{updateProjection,Projection-method}
\alias{updateProjection,Projection-method}
\title{Updates the coordinate data stored in this object.}
\usage{
\S4method{updateProjection}{Projection}(object, data)
}
\arguments{
\item{object}{Projection object}

\item{data}{New data to be stored in the object}
}
\value{
Updated Projection object.
}
\description{
Updates the coordinate data stored in this object.
}
