% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{versionCheck}
\alias{versionCheck}
\title{Check's the version of the FastProject object and displays error if necessary}
\usage{
versionCheck(object)
}
\arguments{
\item{object}{FastProject object}
}
\description{
Check's the version of the FastProject object and displays error if necessary
}
