% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProjectOutput.R
\docType{methods}
\name{viewResults,FastProjectOutput-method}
\alias{viewResults,FastProjectOutput-method}
\alias{viewResults}
\title{View results of analysis without saving output object}
\usage{
\S4method{viewResults}{FastProjectOutput}(object)
}
\arguments{
\item{object}{FastProjectOutput object}
}
\value{
None
}
\description{
launch a local server to explore the results with a browser.
}
\examples{
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

# create housekeeping genes
hkg <- paste0("gene",sample(1:500, 50))

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createUserGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- FastProject(data = expMat,
                     housekeeping = hkg,
                     signatures = sigs)

## Analyze requires actual non-random data to run properly
\dontrun{
fp.out <- Analyze(fp)
viewResults(fp.out)
}
}
