% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\docType{methods}
\name{analyze,FastProject-method}
\alias{analyze,FastProject-method}
\alias{analyze}
\title{Main entry point for running FastProject Analysis}
\usage{
\S4method{analyze}{FastProject}(object, BPPARAM = NULL)
}
\arguments{
\item{object}{FastProject object}

\item{BPPARAM}{a parallelization backend to use for the analysis}
}
\value{
FastProject object
}
\description{
The main analysis function. Runs the entire FastProject analysis pipeline
and returns a FastProject object populated with the result,
}
\examples{
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

# choose housekeeping genes
hkg <- housekeeping$default

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- FastProject(data = expMat,
                     housekeeping = hkg,
                     signatures = sigs)

## analyze requires actual non-random data to run properly
\dontrun{
bp <- BiocParallel::SerialParam()
fp.out <- analyze(fp, BPPARAM=bp)
}
}
