% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{analyze,Vision-method}
\alias{analyze,Vision-method}
\alias{analyze}
\title{Main entry point for running VISION Analysis}
\usage{
\S4method{analyze}{Vision}(object)
}
\arguments{
\item{object}{VISION object}
}
\value{
VISION object
}
\description{
The main analysis function. Runs the entire VISION analysis pipeline
and returns a VISION object populated with the result,
}
\examples{
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- Vision(data = expMat,
             signatures = sigs)

## analyze requires actual non-random data to run properly
\dontrun{
fp.out <- analyze(fp)
}
}
