% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\name{analyze,Vision-method}
\alias{analyze,Vision-method}
\alias{analyze}
\title{Main entry point for running VISION Analysis}
\usage{
\S4method{analyze}{Vision}(object, tree = FALSE, hotspot = FALSE)
}
\arguments{
\item{object}{VISION object}

\item{tree}{wether to use the TREE slot of the object to calculate values}
}
\value{
VISION object
}
\description{
The main analysis function. Runs the entire VISION analysis pipeline
and returns a VISION object populated with the result.
}
\examples{
\dontrun{

vis <- Vision(data = expMat, signatures = sigs)

options(mc.cores=10) # Use 10 cores
vis <- analyze(vis)

}
}
