% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{analyzeProjections}
\alias{analyzeProjections}
\title{Analyze projections}
\usage{
analyzeProjections(object, BPPARAM)
}
\arguments{
\item{object}{the FastProject object}

\item{BPPARAM}{the parallelization backend to use}
}
\value{
the FastProject object with values set for the analysis results
}
\description{
This is the main analysis function. For each filtered dataset, a set of
different projection onto low-dimensional space are computed, and the
consistency of the resulting space with the signature scores is computed
to find signals that are captured succesfully by the projections.
}
