% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{applyFilters}
\alias{applyFilters}
\title{Applies filters to the inputted expression data (may remove rows)}
\usage{
applyFilters(data, threshold, filterInput)
}
\description{
@param data ExpressionData object
 @param threshold minimum number of samples gene must be detected in to pass
 @param filterInput list of filters to compute
 @return List of the updated ExpressionData object, and the list of filters applied.
 @examples
 expr <- readExprToMatrix("data/expression.txt")
 edata <- ExpressionData(expr)
 filteredData <- applyFilters(eData, 10, c("threshold", "fano"))
}
