% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microclusters.R
\name{applyMicroClustering}
\alias{applyMicroClustering}
\title{Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm}
\usage{
applyMicroClustering(exprData, cellsPerPartition = 100,
  filterInput = "fano", filterThreshold = round(ncol(exprData) * 0.2),
  preserve_clusters = NULL, latentSpace = NULL)
}
\arguments{
\item{exprData}{the expression data matrix}

\item{cellsPerPartition}{control over the minimum number of cells to put into each supercell}

\item{preserve_clusters}{named factor vector denoted cluster boundaries to preserve}

\item{latentSpace}{numeric matrix cells x components}
}
\value{
pooled cells - named list of vectors - cells in each supercell
}
\description{
Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm
}
