% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{applyMicroClustering}
\alias{applyMicroClustering}
\title{Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm}
\usage{
applyMicroClustering(exprData, nomodel = TRUE, hkg = matrix(),
  cellsPerPartition = 100, BPPARAM = bpparam())
}
\arguments{
\item{exprData}{the expression data matrix}

\item{nomodel}{a boolean value indicating whether or not to apply FNR curve weight calculations}

\item{hkg}{a matrix vector of house keeping genes to use as negative control}

\item{cellsPerPartition}{control over the minimum number of cells to put into each supercell}

\item{BPPARAM}{the parallelization backend to use}
}
\value{
a list:
\itemize{
    \item pooled cells - the super cells creates
    \item pools - a list of data matrices of the original cells in each pool
}
}
\description{
Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm
}
