% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{applyMicroClustering}
\alias{applyMicroClustering}
\title{Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm}
\usage{
applyMicroClustering(exprData, hkg = list(), BPPARAM = bpparam())
}
\arguments{
\item{exprData}{the expression data matrix}

\item{hkg}{a list of house keeping genes to use as negative control}

\item{BPPARAM}{the parallelization backend to use}
}
\description{
Cluster single cells so signal is maintained but the sample size and noise
are reduce, using the louvain clustering algorithm
}
