% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyPCA}
\alias{applyPCA}
\title{Performs PCA on data}
\usage{
applyPCA(exprData, maxComponents = 200)
}
\arguments{
\item{exprData}{Expression matrix}

\item{maxComponents}{Maximum number of components to calculate}
}
\value{
PCA data

Variance of each component

Eigenvectors of covariance matrix, aka the variable loadings
}
\description{
Performs PCA on data
}
