% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{applyPermutationWPCA}
\alias{applyPermutationWPCA}
\title{Applies pemutation method to return the most significant components of weighted PCA data}
\usage{
applyPermutationWPCA(expr, weights, components = 50, p_threshold = 0.05,
  verbose = FALSE)
}
\arguments{
\item{expr}{Expression data}

\item{weights}{Weights to apply to each coordinate in data}

\item{components}{Maximum components to calculate. Default is 50.}

\item{p_threshold}{P Value to cutoff components at. Default is .05.}

\item{verbose}{Logical value indicating whether or not a verbose session is being run.}
}
\value{
(list):
\itemize{
    \item wPCA: weighted PCA data
    \item eval: the proortinal variance of each component
    \item evec: the eigenvectors of the weighted covariance matrix
    \item permuteMatrices: the permuted matrices generated as the null distrbution
}
}
\description{
Applies pemutation method to return the most significant components of weighted PCA data
}
\details{
Based on the method proposed by Buja and Eyuboglu (1992), PCA is performed on the data
then a permutation procedure is used to assess the significance of components
}
