% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{batchSigEval}
\alias{batchSigEval}
\title{Different ways to evalutate the signature score}
\usage{
batchSigEval(sigs, sig_score_method, eData, weights)
}
\arguments{
\item{sigs}{list of Signature(s) to be evalauting}

\item{sig_score_method}{either "naive" or "weighted_avg"}

\item{eData}{numeric Matrix Genes x Cells}

\item{weights}{Weight matrix computed through FNR curve}
}
\value{
matrix of signature scores, cells X signatures
}
\description{
EAch method should have the same signature so they can be swapped
}
\details{
Right now, each takes in a wrapped data object and signature object

Specified by FastProject argument (sig_score_method), default = naiveEvalSignature
Evaluate signature scores efficiently in batches
}
