% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{calcSignatureScores}
\alias{calcSignatureScores}
\title{calculate signature scores}
\usage{
calcSignatureScores(object, sigData = object@sigData,
  precomputedData = object@precomputedData,
  sig_score_method = object@sig_score_method,
  min_signature_genes = object@min_signature_genes, BPPARAM = NULL)
}
\arguments{
\item{object}{the FastProject object}

\item{sigData}{a list of Signature objects for which to compute the scores}

\item{precomputedData}{a list of existing cell-signatures}

\item{sig_score_method}{the scoring method to use}

\item{min_signature_genes}{the minimal number of genes that must be present
in both the data and the signature for the signature to be evaluated}

\item{BPPARAM}{the parallelization backend to use for these computations}
}
\value{
the FastProject object, with signature score slots populated
}
\description{
For each signature-cell pair, compute a score that captures the level of
correspondence between the cell and the signature.
To estimate significance of these scores, a set of random gene signatures is
generated to create a null distribution
}
