% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{calcSignatureScores}
\alias{calcSignatureScores}
\title{calculate signature scores}
\usage{
calcSignatureScores(object, BPPARAM = NULL)
}
\arguments{
\item{object}{the FastProject object}

\item{BPPARAM}{the parallelization backend to use for these computations}
}
\value{
the FastProject object, with signature score slots populated
}
\description{
For each signature-cell pair, compute a score that captures the level of
correspondence between the cell and the signature.
To estimate significance of these scores, a set of random gene signatures is
generated to create a null distribution
}
