% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{calcSignatureScores}
\alias{calcSignatureScores}
\title{calculate signature scores}
\usage{
calcSignatureScores(object, sigData = object@sigData,
  metaData = object@metaData, sig_norm_method = object@sig_norm_method,
  sig_score_method = object@sig_score_method)
}
\arguments{
\item{object}{the VISION object}

\item{sigData}{a list of Signature objects for which to compute the scores}

\item{metaData}{a list of existing cell-signatures}

\item{sig_norm_method}{(optional) Method to apply to normalize the expression
matrix before calculating signature scores}

\item{sig_score_method}{the scoring method to use}
}
\value{
the VISION object, with signature score slots populated
}
\description{
For each signature-cell pair, compute a score that captures the level of
correspondence between the cell and the signature.
To estimate significance of these scores, a set of random gene signatures is
generated to create a null distribution
}
