% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{calculatePearsonCorr}
\alias{calculatePearsonCorr}
\title{Compute pearson correlation between signature scores and principle components}
\usage{
calculatePearsonCorr(object)
}
\arguments{
\item{object}{the VISION object}
}
\value{
pearsonCorr numeric matrix N_Signatures x N_PCs
}
\description{
Populations the PCAnnotatorData slot of the VISION object
}
