% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{calculateSignatureBackground}
\alias{calculateSignatureBackground}
\title{calculate background signature scores}
\usage{
calculateSignatureBackground(object, num)
}
\arguments{
\item{object}{the VISION object}

\item{num}{the number of signatures to generate}
}
\value{
A list with two items:

  randomSigs: a list of lists of SignatureScore objects.  Each sub-list
    represents permutation signatures generated for a specific size/balance

  sigAssignments: named factor vector assigning signatures to random background
    groups
}
\description{
For each signature-cell pair, compute a score that captures the level of
correspondence between the cell and the signature.
To estimate significance of these scores, a set of random gene signatures is
generated to create a null distribution
}
