% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{calcWeights}
\alias{calcWeights}
\title{Calculate weights based on the false negative rates, computed using the
provided housekeeping genes}
\usage{
calcWeights(object, nomodel = object@nomodel)
}
\arguments{
\item{object}{the FastProject object}

\item{nomodel}{(optional) if TRUE, no fnr curve calculated and all weights
equal to 1. Else FNR and weights calculated.}
}
\value{
the FastProject object with populated weights slot
}
\description{
Calculate weights based on the false negative rates, computed using the
provided housekeeping genes
}
