% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{clusterCells}
\alias{clusterCells}
\title{Creates an initial clustering of the cells}
\usage{
clusterCells(object)
}
\arguments{
\item{object}{the FastProject object for which to cluster the cells}
}
\value{
the FastProject object modifed as described above
}
\description{
Results of this are stored as a new variabe in the object's metaData
and 'cluster_variable' is populated with its name
}
