% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{clusterSignatures}
\alias{clusterSignatures}
\title{Clusters signatures according to the rank sum}
\usage{
clusterSignatures(sigList, sigMatrix, pvals, k = 10)
}
\arguments{
\item{sigList}{List of signatures}

\item{sigMatrix}{Matrix of signatures scores, NUM_SIGNATURES x NUM_SAMPLES}

\item{k}{Number of clusters to generate}
}
\value{
a list:
\itemize{
    \item Computed: a list of clusters of computed signatures
    \item Precomputed: a list of clusters of precomputed signatures
}
}
\description{
Clusters signatures according to the rank sum
}
