% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{clusterSignatures}
\alias{clusterSignatures}
\title{Clusters signatures according to the rank sum}
\usage{
clusterSignatures(sigMatrix, metaData, pvals, clusterMeta)
}
\arguments{
\item{sigMatrix}{data.frame of signatures scores, NUM_SAMPLES x NUM_SIGNATURES}

\item{metaData}{data.frame of meta-data for cells}

\item{pvals}{the corresponding P-values for each score,
NUM_SIGNATURES x NUM_SAMPLES}
}
\value{
a list:
\itemize{
    \item Computed: a list of clusters of computed signatures
    \item Meta: a list of clusters of meta data signatures
}
}
\description{
Clusters signatures according to the rank sum
}
