% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{clusterSigScores}
\alias{clusterSigScores}
\title{Compute KS Test, for all factor meta data.  One level vs all others}
\usage{
clusterSigScores(object)
}
\arguments{
\item{object}{the FastProject object}
}
\value{
the FastProject object with values set for the analysis results
}
\description{
Compute KS Test, for all factor meta data.  One level vs all others
}
