% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{computeKNNWeights,matrix-method}
\alias{computeKNNWeights,matrix-method}
\title{compute for each vector the weights to apply to it's K nearest neighbors}
\usage{
\S4method{computeKNNWeights}{matrix}(object, K = round(sqrt(nrow(object))))
}
\arguments{
\item{object}{matrix to use for KNN}

\item{K}{Number of neighbors to consider.}
}
\value{
a list of two items:
         indices: matrix, cells X neighbors
             Each row specifies indices of nearest neighbors
         weights: matrix, cells X neighbors
             Corresponding weights to nearest neighbors
}
\description{
compute for each vector the weights to apply to it's K nearest neighbors
}
