% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TreeProjection.R
\docType{methods}
\name{computeKNNWeights,TreeProjection-method}
\alias{computeKNNWeights,TreeProjection-method}
\title{Compute KNN weights based on geodesic distances for TreeProjection objects}
\usage{
\S4method{computeKNNWeights}{TreeProjection}(object, K = 30,
  BPPARAM = bpparam())
}
\arguments{
\item{object}{a TreeProjection object}

\item{K}{the number of nearest neighbors to look at}

\item{BPPARAM}{the parallelization backen to use}
}
\value{
an all-pars distance matrix
}
\description{
Compute KNN weights based on geodesic distances for TreeProjection objects
}
