% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{createNewFP}
\alias{createNewFP}
\title{Creates a new FastProject object with the new expression matrix. Called from
Server.R.}
\usage{
createNewFP(fpParams, nexpr)
}
\arguments{
\item{fpParams}{List of FastProject parameters}

\item{nexpr}{New expression matrix for the new FastProject object}
}
\value{
a new FastProject object
}
\description{
Creates a new FastProject object with the new expression matrix. Called from
Server.R.
}
\examples{
  fp <- FastProject("data/expression_matrix.txt",
                    "data/Gene Name Housekeeping.txt",
                    c("data/sigfile_1.gmt", "data/sigfile_2.txt"),
                    precomputed="data/pre_sigs.txt")
  fpslots <- slotNames(object)
  fpParams <- list()
  for (s in fpslots) {
    fpParams <- c(fpParams, slot(object, s))
  }
  subset_cells <- sample(1:ncol(fp@exprData), 200)
  nexpr <- fp@exprData[,subset_cells]
  names(fpParams) <- fpslots
  nfp <- createNewFP(fpParams, nexpr)
}
