% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{createNewFP}
\alias{createNewFP}
\title{create new FastProject object from a subset of the data in an existing one}
\usage{
createNewFP(fp, subset)
}
\arguments{
\item{fp}{the FastProject object to subset}

\item{subset}{the indices of the samples to keep}
}
\value{
a new FastProject object with the new data and the same analysis
parameters
}
\description{
create new FastProject object from a subset of the data in an existing one
}
