% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{createPools}
\alias{createPools}
\title{create "super-cells" by pooling together single cells}
\usage{
createPools(cl, expr)
}
\arguments{
\item{cl}{cluster association of each cell (list of vector of column names)}

\item{expr}{expression data (genes x cells matrix)}
}
\value{
a matrx of expression data for the pooled cells (genes x pools)
}
\description{
create "super-cells" by pooling together single cells
}
