% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transforms.R
\name{createPools}
\alias{createPools}
\title{create "super-cells" by pooling together single cells}
\usage{
createPools(cl, expr, weights)
}
\arguments{
\item{cl}{cluster association of each cell}

\item{expr}{expression data}

\item{weights}{the weight matrix computed from the FNR curve}
}
\value{
a matrx of expression data for the pooled cells
}
\description{
create "super-cells" by pooling together single cells
}
