% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-FastProject.R
\docType{methods}
\name{FastProject}
\alias{FastProject}
\alias{FastProject,matrix-method}
\alias{FastProject,character-method}
\alias{FastProject,ExpressionSet-method}
\alias{FastProject,SummarizedExperiment-method}
\title{Initializes a new FastProject object.}
\usage{
FastProject(data, ...)

\S4method{FastProject}{matrix}(data, signatures, housekeeping = NULL,
  norm_methods = NULL, precomputed = NULL, nomodel = FALSE,
  projection_genes = c("fano"), lean = FALSE, min_signature_genes = 5,
  weights = NULL, threshold = 0, perm_wPCA = FALSE,
  sig_norm_method = "znorm_rows", sig_score_method = "weighted_avg",
  pool = FALSE, cellsPerPartition = 100)

\S4method{FastProject}{character}(data, ...)

\S4method{FastProject}{ExpressionSet}(data, ...)

\S4method{FastProject}{SummarizedExperiment}(data, ...)
}
\arguments{
\item{data}{expression data - can be one of these: \itemize{
\item numeric matrix
\item path of a file,
\item ExpressionSet object
\item SummerizedExperiment object (or extending classes)
}}

\item{...}{additional arguments}

\item{signatures}{list of file paths to signature files (.gmt or .txt) or a
list of Signature objects}

\item{housekeeping}{vector of gene names}

\item{norm_methods}{normalization methods to be extracted from the scone
object}

\item{precomputed}{data file with precomputed signature scores (.txt), or a
data.frame with meta information. Note that rows should match samples in the
data, and columns should either be factors or numerics.}

\item{nomodel}{if TRUE, no fnr curve calculated and all weights equal to 1.
Else FNR and weights calculated. [Default:FALSE]}

\item{projection_genes}{name of method ('threshold' or 'fano') or list of 
genes to use when computing projections.}

\item{lean}{if TRUE run a lean simulation. Else more robust pipeline
initiated. Default is FALSE}

\item{min_signature_genes}{Minimum number of genes required to compute a
signature}

\item{weights}{Precomputed weights for each coordinate. Normally computed
from the FNR curve.}

\item{threshold}{Threshold to apply for the threshold filter}

\item{perm_wPCA}{If TRUE, apply permutation WPCA to calculate significant
number of PCs. Else not. Default FALSE.}

\item{sig_norm_method}{Method to apply to normalize the expression matrix
before calculating signature scores}

\item{sig_score_method}{Method to apply when calculating signature scores}

\item{pool}{a boolean value indicating whether or not to create supercells}

\item{cellsPerPartition}{the minimum number of cells to put into a cluster}
}
\value{
A FastProject object
}
\description{
Initializes a new FastProject object.
}
\examples{
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

# choose housekeeping genes
hkg <- housekeeping$default

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- FastProject(data = expMat,
                     signatures = sigs,
                     housekeeping = hkg)
}
