% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-FastProject.R
\docType{methods}
\name{FastProject}
\alias{FastProject}
\alias{FastProject,matrixORSparse-method}
\alias{FastProject,data.frame-method}
\alias{FastProject,ExpressionSet-method}
\alias{FastProject,SummarizedExperiment-method}
\title{Initializes a new FastProject object.}
\usage{
FastProject(data, ...)

\S4method{FastProject}{matrixORSparse}(data, signatures, housekeeping = NULL,
  unnormalizedData = NULL, meta = NULL, nomodel = TRUE,
  projection_genes = c("fano"), lean = "auto", min_signature_genes = 5,
  weights = NULL, threshold = 0, perm_wPCA = FALSE,
  sig_norm_method = c("znorm_columns", "none", "znorm_rows",
  "znorm_rows_then_columns", "rank_norm_columns"),
  sig_score_method = c("naive", "weighted_avg"),
  trajectory_method = c("None", "SimplePPT"), pool = "auto",
  cellsPerPartition = 100, name = NULL, cluster_variable = "",
  latentSpace = NULL)

\S4method{FastProject}{data.frame}(data, ...)

\S4method{FastProject}{ExpressionSet}(data, ...)

\S4method{FastProject}{SummarizedExperiment}(data, ...)
}
\arguments{
\item{data}{expression data - can be one of these: \itemize{
\item numeric matrix
\item ExpressionSet object
\item SummzrizedExperiment object (or extending classes)
}}

\item{signatures}{list of file paths to signature files (.gmt or .txt) or a
list of Signature objects}

\item{housekeeping}{vector of gene names}

\item{meta}{data file with cell meta data (.txt), or a
data.frame with meta information. Note that rows should match samples in the
data, and columns should either be factors or numerics.}

\item{nomodel}{if TRUE, no fnr curve calculated and all weights equal to 1.
Else FNR and weights calculated. [Default:TRUE]}

\item{projection_genes}{name of method ('threshold' or 'fano') or list of
genes to use when computing projections.}

\item{lean}{if TRUE omit some projection methods with poor scaling.
Default is 'auto' which enables additional methods if the number of cells
is less than 500.}

\item{min_signature_genes}{Minimum number of genes required to compute a
signature}

\item{weights}{Precomputed weights for each coordinate. Normally computed
from the FNR curve.}

\item{threshold}{Threshold to apply for the threshold filter}

\item{perm_wPCA}{If TRUE, apply permutation WPCA to calculate significant
number of PCs. Else not. Default FALSE.}

\item{sig_norm_method}{Method to apply to normalize the expression matrix
before calculating signature scores. Valid options are:
"znorm_columns" (default), "none", "znorm_rows", "znorm_rows_then_columns",
or "rank_norm_columns"}

\item{sig_score_method}{Method to apply when calculating signature scores.
Either "naive" (default) or "weighted_avg"}

\item{trajectory_method}{Method to use to infer a trajectory.  Either
"None" (default) or "SimplePPT".}

\item{pool}{indicates whether or not to create supercells. Acceptable values
are TRUE, FALSE, or 'auto', the last of which is the default and enables
pooling if there are more than 15000 cells.}

\item{cellsPerPartition}{the minimum number of cells to put into a cluster}

\item{name}{a name for the sample - shown on the output report}

\item{cluster_variable}{variable to use to denote clusters}

\item{latentSpace}{latent space for expression data. Numeric matrix or dataframe
with dimensions CELLS x COMPONENTS}
}
\value{
A FastProject object
}
\description{
Initializes a new FastProject object.
}
\examples{
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

# choose housekeeping genes
hkg <- housekeeping$default

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- FastProject(data = expMat,
                     signatures = sigs,
                     housekeeping = hkg)
}
