% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-FastProject.R
\docType{methods}
\name{FastProject}
\alias{FastProject}
\alias{FastProject,matrix-method}
\alias{FastProject,character-method}
\alias{FastProject,ExpressionSet-method}
\title{Initializes a new FastProject object.}
\usage{
FastProject(data, ...)

\S4method{FastProject}{matrix}(data, signatures, housekeeping,
  norm_methods = NULL, precomputed = NULL, nofilter = FALSE,
  nomodel = FALSE, filters = c("fano"), lean = FALSE, qc = FALSE,
  min_signature_genes = 5, projections = "", weights = NULL,
  threshold = 0, perm_wPCA = FALSE, sig_norm_method = "znorm_rows",
  sig_score_method = "weighted_avg")

\S4method{FastProject}{character}(data, ...)

\S4method{FastProject}{ExpressionSet}(data, ...)
}
\arguments{
\item{data}{expression data - can be numeric matrix, path of a file or
ExpressionSet object}

\item{...}{additional arguments}

\item{signatures}{list of file paths to signature files (.gmt or .txt)}

\item{housekeeping}{file path to housekeeping data file}

\item{norm_methods}{normalization methods to be extracted from the scone
object}

\item{precomputed}{data file with precomputed signature scores (.txt)}

\item{nofilter}{if TRUE, no filter applied; else filters applied.
Default is FALSE}

\item{nomodel}{if TRUE, no fnr curve calculated and all weights equal to 1.
Else FNR and weights calculated.
         Default is TRUE.}

\item{filters}{list of filters to compute}

\item{lean}{if TRUE run a lean simulation. Else more robust pipeline
initiated. Default is FALSE}

\item{qc}{if TRUE calculate QC; else not. Default is FALSE}

\item{min_signature_genes}{Minimum number of genes required to compute a
signature}

\item{projections}{File containing precomputed projections for analysis}

\item{weights}{Precomputed weights for each coordinate. Normally computed
from the FNR curve.}

\item{threshold}{Threshold to apply for the threshold filter}

\item{perm_wPCA}{If TRUE, apply permutation WPCA to calculate significant
number of PCs. Else not. Default FALSE.}

\item{sig_norm_method}{Method to apply to normalize the expression matrix
before calculating signature scores}

\item{sig_score_method}{Method to apply when calculating signature scores}
}
\value{
A FastProject object
}
\description{
Initializes a new FastProject object.
}
\examples{
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

# create housekeeping genes
hkg <- paste0("gene",sample(1:500, 50))

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createUserGeneSignature(name = paste0("sig",i),
                               sigData = sigData))
})

fp <- FastProject(data = expMat,
                  signatures = sigs,
                  housekeeping = hkg)
}
