% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{fbConsistencyScores}
\alias{fbConsistencyScores}
\title{Evaluates the significance of each protein}
\usage{
fbConsistencyScores(weights, proteinData)
}
\arguments{
\item{weights}{output of computeKNNWeights}

\item{proteinData}{numeric matrix of protein abundance
size is cells x proteins}
}
\value{
dataframe with columns "C", "pValue", and "FDR"
}
\description{
Evaluates the significance of each protein
}
