% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{filterData}
\alias{filterData}
\title{filter data accourding to the provided filters}
\usage{
filterData(object)
}
\arguments{
\item{object}{the FastProject object}
}
\value{
the FastProject object, populated with filtered data
}
\description{
filter data accourding to the provided filters
}
