% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterGenesFano}
\alias{filterGenesFano}
\title{Applies the Fano filter to the input data (may remove rows)}
\usage{
filterGenesFano(data, num_mad = 2)
}
\arguments{
\item{data}{NUM_GENES x NUM_SAMPLES expression matrix}

\item{num_mad}{number of median absolute deviations}
}
\value{
NUM_GENES_PASSED_FANO_FILTER x NUM_SAMPLES filtered expression matrix
}
\description{
Applies the Fano filter to the input data (may remove rows)
}
