% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterGenesThreshold}
\alias{filterGenesThreshold}
\title{Filter genes whose values sum to less than some threshold value (may remove rows)}
\usage{
filterGenesThreshold(data, threshold)
}
\arguments{
\item{data}{(data.frame) expression matrix}

\item{threshold}{(int) threshold value to filter by}
}
\value{
character vector of gene names passing filter
}
\description{
Filter genes whose values sum to less than some threshold value (may remove rows)
}
