% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{generatePermutationNull}
\alias{generatePermutationNull}
\title{Generate random signatures for a null distribution by permuting the data}
\usage{
generatePermutationNull(num, eData, sigSizes)
}
\arguments{
\item{num}{the number of signatures to generate}

\item{eData}{the data to use for the permutations}

\item{sigSizes}{the sizes of the true signatures, to inform selection of the
random signature sizes}
}
\value{
a vector of random Signature objects
}
\description{
Generate random signatures for a null distribution by permuting the data
}
