% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{generatePermutationNull}
\alias{generatePermutationNull}
\title{Generate random signatures for a null distribution by permuting the data}
\usage{
generatePermutationNull(num, eData, sigData)
}
\arguments{
\item{num}{the number of signatures to generate}

\item{eData}{the data to use for the permutations}

\item{sigData}{list of signature objects
random signature sizes}
}
\value{
A list with two items:

  randomSigs: a list of lists of Signature objects.  Each sub-list represents
    permutation signatures generated for a specific size/balance

  sigAssignments: named factor vector assigning signatures to random background
    groups
}
\description{
Generate random signatures for a null distribution by permuting the data
}
