% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{generateProjections}
\alias{generateProjections}
\title{generate projections}
\usage{
generateProjections(object)
}
\arguments{
\item{object}{the VISION object}
}
\value{
the VISION object with values set for the analysis results
}
\description{
Generates 2-dimensional representations of the expression matrix
Populates the 'Projections' slot on the VISION object
}
