% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{generateProjections}
\alias{generateProjections}
\title{generate projections}
\usage{
generateProjections(object, lean = object@lean)
}
\arguments{
\item{object}{the FastProject object}

\item{lean}{if TRUE run a lean simulation. Else more robust pipeline
initiated. Default is FALSE}
}
\value{
the FastProject object with values set for the analysis results
}
\description{
Generates 2-dimensional representations of the expression matrix
Populates the 'Projections' slot on the FastProject object
}
