% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{generateTreeProjections}
\alias{generateTreeProjections}
\title{Genrate projections based on a tree structure learned in high dimensonal space}
\usage{
generateTreeProjections(data, filterName = "", inputProjections,
  permMats = NULL, BPPARAM = bpparam())
}
\arguments{
\item{data}{the data to fit the tree to. Should be lower dimensional than
full data.}

\item{filterName}{the filtered data to use}

\item{inputProjections}{a list of Projection objects. For each Projection, a
corresponding TreeProjection will be created in which the scores are based
on geodesic distances instead of euclidean one}

\item{permMats}{a list of matrices to use as a null distribution for
estimating the significance of the fitted tree. These are generated by the
permutation wPCA algorithm upstream}

\item{BPPARAM}{the backend to use for parallelization for this analysis}
}
\value{
a list:
\itemize{
    \item projections a list of TreeProjection objects
    \item treeScore a score representing the singificance of the fitten tree
    return(list(projections = output, treeScore = hdTree$zscore))
}
}
\description{
Genrate projections based on a tree structure learned in high dimensonal space
}
