% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\docType{methods}
\name{getNormalizedCopy,ExpressionData-method}
\alias{getNormalizedCopy,ExpressionData-method}
\title{Calculates the specified normalized data matrix}
\usage{
\S4method{getNormalizedCopy}{ExpressionData}(object, func)
}
\description{
@param object ExpressionData object
 @param func normalization method to apply
 @return Normalized data matrix according to function specified.  
 @example 
 eData <- ExpressionData(expr)
 ne <- getNormalizedCopy(eData, "znorm_rows")
}
