% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Cluster.R
\docType{methods}
\name{initialize,Cluster-method}
\alias{initialize,Cluster-method}
\title{Wrapper class for a particular cluster.
Should be called using the `new` syntax.
Maps a cluster type to the the resulting cluster data.}
\usage{
\S4method{initialize}{Cluster}(.Object, method, param, centers, data)
}
\arguments{
\item{.Object}{An object.}

\item{method}{clustering method used}

\item{param}{clusterng parameter used}

\item{centers}{cluster centroid}

\item{data}{cluster-assocated data}
}
\value{
a Cluster object
}
\description{
Wrapper class for a particular cluster.
Should be called using the `new` syntax.
Maps a cluster type to the the resulting cluster data.
}
