% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ExpressionData.R
\docType{methods}
\name{initialize,ExpressionData-method}
\alias{initialize,ExpressionData-method}
\title{Initializes an ExpressionData object
should not be called directly, use the `new` syntax}
\usage{
\S4method{initialize}{ExpressionData}(.Object, data, ...)
}
\arguments{
\item{.Object}{an object}

\item{data}{expression data matrix}

\item{...}{additional arguments}
}
\value{
ExpressionData object
}
\description{
Initializes an ExpressionData object
should not be called directly, use the `new` syntax
}
