% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProjectOutput.R
\docType{methods}
\name{initialize,FastProjectOutput-method}
\alias{initialize,FastProjectOutput-method}
\title{Initialize a new FastProjectOuptput object}
\usage{
\S4method{initialize}{FastProjectOutput}(.Object, eData, filterModuleList,
  sigMatrix, sigList, fpParams, pools)
}
\arguments{
\item{eData}{ExpressionData matrix}

\item{filterModuleList}{List of FilterModuleData objects}

\item{sigList}{List of Signatures}

\item{fpParams}{List of the FastProject parameters that were used to generate this object}

\item{sigMatirx}{Matrix of signature scores (NUM_SIGS x NUM_CELLS)}
}
\value{
FastProjectOutput object.
}
\description{
Initialize a new FastProjectOuptput object
}
