% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\docType{methods}
\name{initialize,Signature-method}
\alias{initialize,Signature-method}
\title{Initialize a new Signature object.
Should not be called directly, instead use the `new` syntax}
\usage{
\S4method{initialize}{Signature}(.Object, sigDict, name, source,
  metaData = "", isPrecomputed = FALSE, isFactor = FALSE, cluster = 0)
}
\arguments{
\item{.Object}{an object}

\item{sigDict}{Named list of signs for each gene in the signature}

\item{name}{Name of the signature}

\item{source}{File from which this signature was read from}

\item{metaData}{Metadata pertinent to signature}

\item{isPrecomputed}{If TRUE indicates that this signature was precomputed.
Else not precomputed. Default is FALSE.}

\item{isFactor}{If TRUE indicates that htis signature is a Factor, else not
a factor. Default is FALSE.}

\item{cluster}{Number representing which cluster this signature is a part of.
Default is 0.}
}
\value{
Signature object
}
\description{
Initialize a new Signature object.
Should not be called directly, instead use the `new` syntax
}
