% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{lcaBasedTreeKNN}
\alias{lcaBasedTreeKNN}
\title{Add custom tree based neighbor and weights to a hotspot object}
\usage{
lcaBasedTreeKNN(tree, minSize = 20)
}
\description{
@param tree object of class phylo
 @param the hotspot object to add the nw to
 @param minSize the minimum number of neighbors of the node
 @return the hotspot object
}
