% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProject.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize the data}
\usage{
normalizeData(object)
}
\arguments{
\item{object}{the FastPrpject object}
}
\value{
the FastProject object with normalized data matrices
}
\description{
Normalize the data
}
