% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{poolCells}
\alias{poolCells}
\title{create micro-clusters that reduce noise and complexity while maintaining
the overall signal in the data}
\usage{
poolCells(object, BPPARAM = SerialParam())
}
\arguments{
\item{object}{the FastProject object to cluster the cells of}

\item{BPPARAM}{the parallelization backend to use}
}
\value{
the FastProject with pooled cells
}
\description{
create micro-clusters that reduce noise and complexity while maintaining
the overall signal in the data
}
