% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{processSignatures}
\alias{processSignatures}
\title{Processes signatures filtering unmatched genes}
\usage{
processSignatures(sigData, expressionGenes, minSignatureGenes)
}
\arguments{
\item{sigData}{list of signature objects}

\item{expressionGenes}{list of gene identifiers in the expression matrix}

\item{processedSigData}{resulting signature objects}
}
\description{
Processes signatures filtering unmatched genes
}
