% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ProjectionData.R
\name{ProjectionData}
\alias{ProjectionData}
\title{Wrapper for storing all relevant information for a given projection.}
\usage{
ProjectionData(projections = NULL, sigProjMatrix, pMatrix, sigClusters,
  emp_pMatrix)
}
\arguments{
\item{projections}{List of Projection objects to be stored}

\item{sigProjMatrix}{Matrix storing the median consistency score per projection, signature pair}

\item{pMatrix}{Matrix storing the p values for each projection, signature pair}

\item{sigClusters}{the signature clusters}
}
\value{
ProjectionData object
}
\description{
Stores a list of Projection objects
Also stores clusters, a signature to projection matrix
Initializes a ProjectionData object for neatly storing all relevant data to a given projection section
}
