% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{read_10x_h5}
\alias{read_10x_h5}
\title{Read 10x HDF5 Output}
\usage{
read_10x_h5(h5_file, ensToSymbol = TRUE)
}
\arguments{
\item{h5_file}{path to h5 file}

\item{ensToSymbol}{bool denoting whether or not to perform label conversion}
}
\value{
sparse count matrix with appropriate row/column names
}
\description{
Loads 10x output counts and converts expression to gene symbols
}
\details{
This version uses the h5 file produced by "cellranger count"

This file is typically in a folder that looks like:

\code{outs/filtered_gene_bc_matrices_h5.h5}
}
