% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readExprToMatrix}
\alias{readExprToMatrix}
\title{Reads in an expression matrix from a file path to a matrix}
\usage{
readExprToMatrix(filename, delimiter = "\\t")
}
\arguments{
\item{filename}{path to the file to be read in}

\item{delimiter}{delimiter to use while reading in the file. Default is "\\t"}
}
\value{
Matrix read in from filename.
}
\description{
Reads in an expression matrix from a file path to a matrix
}
