% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readPrecomputed}
\alias{readPrecomputed}
\title{Reads precomputed signature values form a tab-delimited text file}
\usage{
readPrecomputed(filename, sampleLabels, delimiter = "\\t")
}
\description{
First row of the file contains sample labels that the signatures correspond with
}
\details{
Each subsequent row contains a signature name in the first column,
 followed by the signature type (either 'numerical' or 'factor')
 followed by the signature values, one for each sample label in the file

@param  filename Filename to read the precomputed signatures from
 @param sampleLables List of labels for which we want the signature scores
 @param delimiter Delimiter to use when reading in file. Default is tab.
 @returns List of Signature data types
 @examples
 expr <- readExprToMatrix("expression.txt")
 samples <- colnames(expr)
 precomputed <- readPrecomputed("precomputed_sigs.txt", samples)
}
