% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projections.R
\name{registerMethods}
\alias{registerMethods}
\title{Registers the projection methods to be used}
\usage{
registerMethods(lean = FALSE)
}
\arguments{
\item{lean}{If FALSE, all projections applied; else a subset of essential ones are applied. Default is FALSE.}
}
\value{
List of projection methods to be applied.
}
\description{
Registers the projection methods to be used
}
