% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProjectOutput.R
\docType{methods}
\name{saveFPOutAndViewResults,FastProjectOutput-method}
\alias{saveFPOutAndViewResults,FastProjectOutput-method}
\alias{saveFPOutAndViewResults}
\title{Save the FastProjectOutput object as an .RDS file and view the results on a
localhost}
\usage{
\S4method{saveFPOutAndViewResults}{FastProjectOutput}(fpout)
}
\arguments{
\item{fpout}{FastProjectOutput object}
}
\value{
None
}
\description{
Save the results object as an RDS file for future use, and launch a local
server to explore the results with a browser.
}
\examples{
\dontrun{
fpout <- Analyze(fp)
saveFPOutAndViewResults(fpout)
}
}
