% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{ServerExpression}
\alias{ServerExpression}
\title{Wrapper class for gene expression object for JSON.}
\usage{
ServerExpression(data, sample_labels, gene_labels)
}
\arguments{
\item{data}{numeric matrix}

\item{sample_labels}{Labels of samples in expression data}

\item{gene_labels}{Lables of genes in expression data}
}
\value{
ServerExpression object
}
\description{
Wrapper class for gene expression object for JSON.
}
