% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{ServerSigProjMatrix}
\alias{ServerSigProjMatrix}
\title{Wrapper class for Signature Projection Matrix}
\usage{
ServerSigProjMatrix(zscores, pvals, proj_labels, sig_labels)
}
\arguments{
\item{zscores}{numeric matrix}

\item{pvals}{numeric matrix}

\item{proj_labels}{Projection names}

\item{sig_labels}{Names of signatures}
}
\value{
ServerSigProjMatrix object
}
\description{
Wrapper class for Signature Projection Matrix
}
