% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{Signature}
\alias{Signature}
\title{Initialize a new Signature object.
Should not be called directly, instead use the `new` syntax}
\usage{
Signature(sigDict, name, source, metaData = "", isMeta = FALSE,
  isFactor = FALSE)
}
\arguments{
\item{sigDict}{Named list of signs for each gene in the signature}

\item{name}{Name of the signature}

\item{source}{File from which this signature was read from}

\item{metaData}{Metadata pertinent to signature}

\item{isMeta}{If TRUE indicates that this signature was derived from
meta data. Default is FALSE.}

\item{isFactor}{If TRUE indicates that this signature is a Factor, else not
a factor. Default is FALSE.}
}
\value{
Signature object
}
\description{
Initialize a new Signature object.
Should not be called directly, instead use the `new` syntax
}
