% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SignatureScores.R
\name{SignatureScores}
\alias{SignatureScores}
\title{Initializes a new SignatureScores object
Should not be called directly, instead use the `new` syntax}
\usage{
SignatureScores(scores, name, isFactor, isPrecomputed, numGenes)
}
\arguments{
\item{scores}{Signature scores}

\item{name}{Name of signature}

\item{isFactor}{Indicates whether or not this is a factor signature}

\item{isPrecomputed}{Indicates whether or not this score was precomputed}

\item{numGenes}{The number of genes used to calculate the score}
}
\value{
New SignatureScores object
}
\description{
Initializes a new SignatureScores object
Should not be called directly, instead use the `new` syntax
}
