% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{signatureToJSON}
\alias{signatureToJSON}
\title{Converts Signature object to JSON}
\usage{
signatureToJSON(sig)
}
\arguments{
\item{sig}{Signature object}
}
\value{
JSON formatted Signature object.
}
\description{
Converts Signature object to JSON
}
