% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{sigProjMatrixToJSON}
\alias{sigProjMatrixToJSON}
\title{Converts a sigProjMatrix from a FastProjectOutput Object to a JSON object}
\usage{
sigProjMatrixToJSON(sigpm, sigs)
}
\arguments{
\item{sigpm}{SigProjMatrix}

\item{sigs}{Signatures to subset form sigpm}
}
\value{
Subsetted sigProjMatirx converted to JSON
}
\description{
Converts a sigProjMatrix from a FastProjectOutput Object to a JSON object
}
