% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{sigScoresToJSON}
\alias{sigScoresToJSON}
\title{Converts row of signatures score matrix to JSON}
\usage{
sigScoresToJSON(names, values)
}
\arguments{
\item{names}{character vector of labels for signatures}

\item{values}{numeric vector for signature values}
}
\value{
Signature scores list to JSON, with names of each entry that of the list names
}
\description{
Converts row of signatures score matrix to JSON
}
