% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{sigScoresToJSON}
\alias{sigScoresToJSON}
\title{Converts row of sigantures score matrix to JSON}
\usage{
sigScoresToJSON(names, values)
}
\arguments{
\item{ss}{single-column dataframe with scores for a single signature}
}
\value{
Signature scores list to JSON, with names of each entry that of the list names
}
\description{
Converts row of sigantures score matrix to JSON
}
