% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{sigsVsProjection_n}
\alias{sigsVsProjection_n}
\title{Evaluates the significance of each numeric signature vs. a
single projections weights}
\usage{
sigsVsProjection_n(sigScores, randomSigData, normExpr, weights, cells = NULL)
}
\arguments{
\item{sigScores}{numeric matrix of signature scores
size is cells x signatures}

\item{randomSigData}{A list with three items:

  randomSigs: a list of signature score matrices.  Each list item
    represents permutation signatures generated for a specific size/balance,
    and is a numeric matrix of size cells X signatures

  sigAssignments: named factor vector assigning signatures to random background
    groups

  randomSigAssignments: named factor vector assigning signatures to random background
    groups

obtained from calling generatePermutationNull}

\item{normExpr}{NormData object used to generate random background sigs}

\item{weights}{numeric matrix of dimension N_SAMPLES x N_SAMPLES}

\item{cells}{list of cell names.  Subsets anlysis to provided cells.
If omitted, use all cells.}
}
\value{
list:
\itemize{
    \item consistency: consistency scores
    \item pvals: pvalues
    \item emppvals: empirical pvalues
}
}
\description{
Evaluates the significance of each numeric signature vs. a
single projections weights
}
