% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{sigsVsProjection_n}
\alias{sigsVsProjection_n}
\title{Evaluates the significance of each numeric signature vs. a
single projections weights}
\usage{
sigsVsProjection_n(sigData, sigScoreMatrix, randomSigData, randomSigScoreMatrix,
  weights)
}
\arguments{
\item{sigData}{a list of SignatureScores objects representing input user
signatures}

\item{sigScoreMatrix}{a numeric matrix (N_SAMPLES x N_SIGNATURES) containing
the signature scores for every (signature, sample)}

\item{randomSigData}{a list of SignatureScores objects representing a suitable
background distribution to go with `sigData`}

\item{randomSigScoreMatrix}{a numeric matrix (N_SAMPLES x N_SIGNATURES) containing
the signature scores associated wiht signatures in randomSigData}

\item{weights}{numeric matrix of dimension N_SAMPLES x N_SAMPLES}
}
\value{
list:
\itemize{
    \item consistency: consistency scores
    \item pvals: pvalues
}
}
\description{
Evaluates the significance of each numeric signature vs. a
single projections weights
}
