% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Signature.R
\name{sigsVsProjections}
\alias{sigsVsProjections}
\title{Evaluates the significance of each signature vs. each projection.}
\usage{
sigsVsProjections(projections, sigScoresData, randomSigData,
  NEIGHBORHOOD_SIZE = 0.33, BPPARAM = bpparam())
}
\arguments{
\item{projections}{Maps projections to their spatial coordinates for each sample}

\item{sigScoresData:}{List of SignatureScores Object, mapping signature names to their value at each coordinate}

\item{randomSigData:}{List of SignatureScores Object, mapping randomly generated signatures to scores to be compared with the real signature scores.}
}
\value{
list:
\itemize{
    \item sigNames: labels for the signatures (rows in output matrices)
    \item projNames: labels for the projections (columns in output matrices)
    \item sigProbMatrix: the matrix of signature-projection consistency scores
    \item pVals: pvalues for the scores
}
}
\description{
Evaluates the significance of each signature vs. each projection.
}
