% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubSample.R
\name{splitSamples}
\alias{splitSamples}
\title{This module handles the splitting and joining of the data if sub-sampling is enabled}
\usage{
splitSamples(data, sampleSize)
}
\description{
This module handles the splitting and joining of the data if sub-sampling is enabled
}
