% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{versionCheck}
\alias{versionCheck}
\title{Checks the version of the Vision object and displays error if necessary}
\usage{
versionCheck(object)
}
\arguments{
\item{object}{Vision object}
}
\description{
Checks the version of the Vision object and displays error if necessary
}
