% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastProjectOutput.R
\docType{methods}
\name{viewResults,FastProjectOutput-method}
\alias{viewResults,FastProjectOutput-method}
\title{View results of analysis without saving output object}
\usage{
\S4method{viewResults}{FastProjectOutput}(object)
}
\arguments{
\item{object}{FastProjectOutput object}
}
\description{
launch a local server to explore the results with a browser.
}
\examples{
\dontrun{
fpout <- Analyze(fp)
viewResults(fpout)
}
}
