% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Vision.R
\docType{methods}
\name{viewResults,Vision-method}
\alias{viewResults,Vision-method}
\alias{viewResults}
\alias{viewResults,character-method}
\title{View results of analysis}
\usage{
\S4method{viewResults}{Vision}(object, port = NULL, host = NULL,
  browser = TRUE, name = NULL)

\S4method{viewResults}{character}(object, port = NULL, host = NULL,
  browser = TRUE, name = NULL)
}
\arguments{
\item{object}{VISION object or path to an RDS file containing such an
object (saved using saveAndViewResults, or directly using saveRDS)}

\item{port}{The port on which to serve the output viewer.  If omitted, a
random port between 8000 and 9999 is chosen.}

\item{host}{The host used to serve the output viewer. If omitted, "127.0.0.1"
is used.}

\item{browser}{Whether or not to launch the browser automatically (default=TRUE)}

\item{name}{Name for the sample - is shown at the top of the output report}
}
\value{
None
}
\description{
launch a local server to explore the results with a browser.
}
\examples{
\dontrun{

vis <- Vision(data = expMat, signatures = sigs)

options(mc.cores=10) # Use 10 cores
vis <- analyze(vis)

saveRDS(vis, 'vision_results.rds') # (Optional) Save results

vis <- viewResults(vis) # Launches dynamic output report
}
}
