% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigScoreMethods.R
\name{weightedEvalSignature}
\alias{weightedEvalSignature}
\title{Evaluate signature with weights computed from FNR curve.}
\usage{
weightedEvalSignature(exprData, sig, weights, min_signature_genes)
}
\arguments{
\item{exprData}{numeric Matrix Genes x Cells}

\item{sig}{Signature to be evalauting}

\item{weights}{Weight matrix computed through FNR curve}

\item{min_signature_genes}{Minimum number of genes to evaluate}
}
\value{
SignatureScore object
}
\description{
Evaluate signature with weights computed from FNR curve.
}
